% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Empirical.R
\name{Empirical}
\alias{Empirical}
\alias{pempirical}
\alias{dempirical}
\alias{qempirical}
\alias{rempirical}
\alias{mean.Empirical}
\alias{variance.Empirical}
\alias{skewness.Empirical}
\alias{quantile.Empirical}
\title{Create an Empirical distribution}
\usage{
Empirical(x)

pempirical(q, y, lower.tail = TRUE, log.p = FALSE, na.rm = TRUE)

dempirical(x, y, log = FALSE, method = "hist", ...)

qempirical(p, y, lower.tail = TRUE, log.p = FALSE, na.rm = TRUE, ...)

rempirical(n, y, na.rm = TRUE)

\method{mean}{Empirical}(x, ...)

\method{variance}{Empirical}(x, ...)

\method{skewness}{Empirical}(x, type = 1L, ...)

\method{quantile}{Empirical}(x, probs, drop = TRUE, elementwise = NULL, ...)
}
\arguments{
\item{x}{an object of class \code{Empirical} (see [Empirical()]).}

\item{q}{vector of quantiles.}

\item{y}{vector of observations of the empirical distribution with two or more non-missing finite values.}

\item{lower.tail}{logical; if \code{TRUE} (default), probabilities are
\code{P[X <= x]} otherwise, \code{P[X > x]}.
or \code{"density"}.}

\item{na.rm}{logical evaluating to \code{TRUE} or \code{FALSE} indicating whether
\code{NA} values should be stripped before the computation
proceeds.}

\item{log, log.p}{logical; if \code{TRUE}, probabilities \code{p} are given as \code{log(p)}.}

\item{method}{character; the method to calculate the empirical density. Either \code{"hist"} (default)}

\item{...}{Arguments to be passed to \code{\link[stats]{qnorm}}.
Unevaluated arguments will generate a warning to catch mispellings or other
possible errors.}

\item{p}{vector of probabilities.}

\item{n}{The number of samples to draw. Defaults to `1L`.}

\item{type}{integer between \code{1L} and \code{3L} (default) selecting one of three
algorithms. See Details for more information.}

\item{probs}{A vector of probabilities.}

\item{drop}{logical. Should the result be simplified to a vector if possible?}

\item{elementwise}{logical. Should each distribution in \code{x} be evaluated
at all elements of \code{probs} (\code{elementwise = FALSE}, yielding a matrix)?
Or, if \code{x} and \code{probs} have the same length, should the evaluation be
done element by element (\code{elementwise = TRUE}, yielding a vector)? The
default of \code{NULL} means that \code{elementwise = TRUE} is used if the
lengths match and otherwise \code{elementwise = FALSE} is used.}
}
\value{
An `Empirical` object.

In case of a single distribution object, either a numeric
  vector of length `probs` (if `drop = TRUE`, default) or a `matrix` with
  `length(probs)` columns (if `drop = FALSE`). In case of a vectorized
  distribution object, a matrix with `length(probs)` columns containing all
  possible combinations.
}
\description{
An empirical distribution consists of a series of \code{N} observations
out of a typically unknown distribution, i.e., a random sample 'X'.

Please see the documentation of [Empirical()] for some properties
of the Empirical distribution, as well as extensive examples
showing to how calculate p-values and confidence intervals.
`quantile()`
}
\details{
The creation function [Empirical()] allows for a variety of different objects
  as main input \code{x}.

  * Vector: Assumes that the vector contains a series of observations from one
    empirical distribution.

  * List (named or unnamed) of vectors: Each element in the list describes
    one empirical distribution defined by the numeric values in each of the vectors.

  * Matrix or data.frame: Each row corresponds to one empirical distribution, whilst
    the columns contain the individual observations.

  Missing values are allowed, however, each distribution requires at least two
  finite observations (\code{-Inf}/\code{Inf} is replaced by \code{NA}).

    **Support**: \eqn{R}, the set of all real numbers

    **Mean**: \deqn{\bar{x} = \frac{1}{N} \sum_{i=1}^{N} x_i}{\bar{x} = 1 / N * sum(x)}

    **Variance**: \deqn{\frac{1}{N - 1} \sum_{i=1}^{N} (x_i - \bar{x})}{1 / (N - 1) * sum((x - \bar{x})^2)}

    **Skewness**:

    \eqn{S_1 = \sqrt{N} \frac{\sum_{i=1}^N (x_i - \bar{x})^3}{\sqrt{\big(\sum_{i=1}^N (x_i - \bar{x})^2\big)^3}}}

    \eqn{S_2 = \frac{\sqrt{N * (N - 1)}}{(N - 2)} S_1} (only defined for \eqn{N > 2})

    \eqn{S_3 = \sqrt{(1 - \frac{1}{N})^3} * S_1} (default)

    For more details about the different types of sample skewness see Joanes and Gill (1998).

    **Kurtosis**:

    \eqn{K_1 = N * \frac{\sum_{i=1}^N (x_i - \bar{x})^4}{\big(\sum_{i=1}^N (x_i - \bar{x})^2\big)^2} - 3}

    \eqn{K_2 = \frac{(N + 1) * K_1 + 6) * (N - 1)}{(N - 2) * (N - 3)}} (only defined for \eqn{N > 2})

    \eqn{K_3 = \big(1 - \frac{1}{N}\big)^2 * (K_1 + 3) - 3} (default)

    For more details about the different types of sample kurtosis see Joanes and Gill (1998).

    **TODO(RETO)**: Add empirical distribution function information (step-function 1/N)

    **Probability density function (p.d.f)**:

This function returns the same values that you get from a Z-table. Note
`quantile()` is the inverse of `cdf()`. Please see the documentation of
[Empirical()] for some properties of the Empirical distribution, as well as
extensive examples showing to how calculate p-values and confidence
intervals.
}
\examples{

require("distributions3")
set.seed(28)

X <- Empirical(rnorm(50))
X

mean(X)
variance(X)
skewness(X)
kurtosis(X)

random(X, 10)

pdf(X, 2)
log_pdf(X, 2)

cdf(X, 4)
quantile(X, 0.7)

### example: allowed types/classes of input arguments

## Single vector (will be coerced to numeric)
Y1  <- rnorm(3, mean = -10)
d1 <- Empirical(Y1)
d1
mean(d1)

## Unnamed list of vectors
Y2 <- list(as.character(rnorm(3, mean = -10)),
           runif(6),
           rpois(4, lambda = 15))
d2 <- Empirical(Y2)
d2
mean(d2)

## Named list of vectors
Y3 <- list("Normal"  = as.character(rnorm(3, mean = -10)),
           "Uniform" = runif(6),
           "Poisson" = rpois(4, lambda = 15))
d3 <- Empirical(Y3)
d3
mean(d3)

## Matrix or data.frame
Y4 <- matrix(rnorm(20), ncol = 5, dimnames = list(sprintf("D_\%d", 1:4), sprintf("obs_\%d", 1:5)))
d4 <- Empirical(Y4)
d4
d5 <- Empirical(as.data.frame(Y4))
d5

}
\references{
Joanes DN and Gill CA (1998). \dQuote{Comparing Measures of
Sample Skewness and Kurtosis.} \emph{Journal of the Royal Statistical
Society D}, \bold{47}(1), 183--189. \doi{10.1111/1467-9884.00122}
}
\seealso{
Other Empirical distribution: 
\code{\link{cdf.Empirical}()}
}
\concept{Empirical distribution}
\concept{distributions}
