% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqrplot.R
\docType{data}
\name{geom_qqr_point}
\alias{geom_qqr_point}
\alias{GeomQqrPoints}
\alias{stat_qqr_range}
\alias{StatQqrRange}
\alias{geom_qqr_range}
\alias{GeomQqrRange}
\alias{stat_qqr_ref}
\alias{StatQqrRef}
\alias{geom_qqr_ref}
\alias{GeomQqrRef}
\alias{stat_qqr_confint}
\alias{StatQqrConfint}
\alias{geom_qqr_confint}
\alias{GeomQqrConfint}
\title{\code{geom_*} and \code{stat_*} for Producing Quantile Residual Q-Q Plots with `ggplot2`}
\format{
An object of class \code{GeomQqrConfint} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 6.
}
\usage{
geom_qqr_point(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)

stat_qqr_range(
  mapping = NULL,
  data = NULL,
  geom = "qqr_range",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)

geom_qqr_range(
  mapping = NULL,
  data = NULL,
  stat = "qqr_range",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)

stat_qqr_ref(
  mapping = NULL,
  data = NULL,
  geom = "qqr_ref",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  detrend = FALSE,
  identity = TRUE,
  probs = c(0.25, 0.75),
  trafo = qnorm,
  ...
)

geom_qqr_ref(
  mapping = NULL,
  data = NULL,
  stat = "qqr_ref",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  detrend = FALSE,
  identity = TRUE,
  probs = c(0.25, 0.75),
  trafo = qnorm,
  ...
)

geom_qqr_confint(
  mapping = NULL,
  data = NULL,
  stat = "qqr_confint",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  xlim = NULL,
  n = 101,
  detrend = FALSE,
  identity = TRUE,
  probs = c(0.25, 0.75),
  trafo = qnorm,
  style = c("polygon", "line"),
  ...
)

GeomQqrConfint
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}} or
\code{\link[ggplot2:aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{stat}{The statistical transformation to use on the data for this
layer, as a string.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{geom}{The geometric object to use display the data}

\item{detrend}{Fix description.}

\item{identity}{logical, should the identity line be plotted or a theoretical line
which passes through \code{probs} quantiles computed by \code{trafo}.}

\item{probs}{numeric vector of length two, representing probabilities of reference
line used in \code{trafo}.}

\item{trafo}{function for calculating reference line through first and third
quartile of theoretical distribution (default: Gaussian \code{qnorm}).}

\item{xlim}{Fix description.}

\item{n}{Fix description.}

\item{style}{Fix description.}
}
\description{
Various \code{geom_*} and \code{stat_*} used within
\code{\link[ggplot2]{autoplot}} for producing quantile residual Q-Q plots.
}
\examples{
require("ggplot2")
## Fit model
data("CrabSatellites", package = "countreg")
m1_pois <- glm(satellites ~ width + color, data = CrabSatellites, family = poisson)
m2_pois <- glm(satellites ~ color, data = CrabSatellites, family = poisson)

## Compute qqrplot
q1 <- qqrplot(m1_pois, plot = FALSE)
q2 <- qqrplot(m2_pois, plot = FALSE)

d <- c(q1, q2) 

## Get label names
xlab <- unique(attr(d, "xlab"))
ylab <- unique(attr(d, "ylab"))
main <- attr(d, "main")
main <- make.names(main, unique = TRUE)
d$group <- factor(d$group, labels = main)

## Polygon CI around identity line used as reference 
gg1 <- ggplot(data = d, aes(x, y, na.rm = TRUE)) + 
  geom_qqr_ref() + 
  geom_qqr_confint(fill = "red") + 
  geom_qqr_point() + 
  geom_qqr_range(
    aes(
      x = x_rg_lwr, 
      ymin = y_rg_lwr, 
      ymax = y_rg_upr,
      group = group
    )
  ) + 
  xlab(xlab) + ylab(ylab)

gg1
gg1 + facet_wrap(~group)

## Polygon CI around robust reference line
gg2 <- ggplot(data = d, aes(x, y, na.rm = TRUE)) + 
  geom_qqr_ref(identity = FALSE, trafo = attr(d, "trafo")) + 
  geom_qqr_confint(identity = FALSE, trafo = attr(d, "trafo"), style = "line") + 
  geom_qqr_point() + 
  geom_qqr_range(
    aes(
      x = x_rg_lwr, 
      ymin = y_rg_lwr, 
      ymax = y_rg_upr,
      group = group
    )
  ) + 
  xlab(xlab) + ylab(ylab)

gg2
gg2 + facet_wrap(~group)

}
\keyword{datasets}
