% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hasS3method}
\alias{hasS3method}
\title{Helper functions to check available support for some S3methods}
\usage{
hasS3method(method, classes)
}
\arguments{
\item{methods}{character, name of the method (e.g., is_continuous, print, ...)}
}
\value{
Returns TRUE if the method exists for one of the given classes, else FALSE.
}
\description{
Helper functions to check available support for some S3methods
}
