% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitresiduals.R
\name{pitresiduals}
\alias{pitresiduals}
\alias{pitresiduals.default}
\title{Probability Integral Transform (PIT) Residuals}
\usage{
pitresiduals(object, ...)

\method{pitresiduals}{default}(
  object,
  newdata = NULL,
  scale = c("uniform", "normal"),
  type = c("random", "quantile"),
  nsim = 1L,
  delta = NULL,
  prob = NULL,
  ...
)
}
\arguments{
\item{object}{an object. For the \code{default} method this needs to be
either a specification of probabilities (vector or 2-dimensional matrix of
probabilities) or an object from which the these can be obtained with
\code{\link{procast}}.}

\item{\dots}{further parameters passed to methods.}

\item{newdata}{optionally, a data frame in which to look for variables with
which to predict. If omitted, the original observations are used.}

\item{scale}{On which scale should the PIT residuals be shown; on the probability scale 
(\code{"uniform"}) or on the normal scale (\code{"normal"}).}

\item{type}{character specifying whether - in the case of discrete response
distributions - randomized quantile residuals or their corresponding
quantiles should be computed.}

\item{nsim}{numeric. The number of simulated randomized quantile residuals
per observation (for \code{type = "numeric"}).}

\item{delta}{numeric. The minimal difference to compute the range of
proabilities corresponding to each observation according to get (randomized)
quantile residuals.  For \code{NULL}, the minimal observed difference in the
resonse divided by \code{5e-6} is used.}

\item{prob}{numeric. The probabilities at which quantile residuals should be
computed (for \code{type = "quantile"}), defaulting to the median
(\code{0.5}).}
}
\value{
A vector or matrix of PIT residuals.
}
\description{
Generic function and methods for computing PIT residuals.
}
\details{
For regression models with a continuous response distribution, PIT residuals
are simply the predictive cumulative distribution (CDF) evaluated at the
observations (Dawid, 1984). For discrete distributions, a random value is drawn
from the range of probabilities corresponding to each observation similar to
the approach of Dunn and Smyth (1996) to gain (randomized) quantile residuals
(\code{\link{qresiduals}}). PIT values have been used under various names, but
to emphasize their similar properties to residuals we follow Warton (2017) and
refer to them as PIT residuals.

With the method \code{pitresiduals}, the PIT residuals can also be transformed
from the probability scale to another distribution scale. Supported quantile
scales are uniformly (\code{uniform}) and normally (\code{normal}) distributed.
Probabilites can be either a vector or a 2-column matrix of probabilities. The
latter offers to either draw \code{"random"} samples from the distribution or
compute corresponding \code{"quantile"}s such as the median etc. For
(randomized) quantile residuals (on the normal scale), as suggested by Dunn and Smyth (1996), 
the \code{scale} must be set to \code{"normal"} or 
\code{\link{qresiduals}} can be called directly.
}
\examples{

## linear regression models (homoscedastic Gaussian response)
m <- lm(dist ~ speed, data = cars)
pitresiduals(m)

}
\references{
Dunn KP, Smyth GK (1996). \dQuote{Randomized Quantile
Residuals.} \emph{Journal of Computational and Graphical Statistics},
\bold{5}, 1--10. \doi{10.2307/1390802}.

Dawid AP (1984). \dQuote{Present Position and Potential Developments: Some
Personal Views: Statistical Theory: The Prequential Approach.} \emph{Journal of
the Royal Statistical Society: Series A (General)}, \bold{147}(2), 278--92.
\doi{10.2307/2981683}.

Warton DI, Thibaut L, Wang YA (2017) \dQuote{The Pit-Trap--a `Model-Free'
Bootstrap Procedure for Inference About Regression Models with Discrete,
Multivariate Responses}. \emph{PLOS ONE}, \bold{12}(7), 1--18.
\doi{10.1371/journal.pone.0181790}.
}
\seealso{
\code{\link[stats]{qnorm}}, \code{\link{qqrplot}}
}
\keyword{regression}
