% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pithist.R
\name{plot.pithist}
\alias{plot.pithist}
\alias{lines.pithist}
\alias{autoplot.pithist}
\title{Plotting a PIT Histogram}
\usage{
\method{plot}{pithist}(
  x,
  single_graph = FALSE,
  style = c("histogram", "lines"),
  confint = TRUE,
  ref = TRUE,
  xlim = c(0, 1),
  ylim = c(0, NA),
  xlab = NULL,
  ylab = NULL,
  main = NULL,
  col = "black",
  fill = adjustcolor("black", alpha.f = 0.2),
  border = "black",
  alpha_min = 0.2,
  lwd = NULL,
  lty = 1,
  axes = TRUE,
  box = TRUE,
  ...
)

\method{lines}{pithist}(
  x,
  confint = FALSE,
  ref = FALSE,
  col = "black",
  fill = adjustcolor("black", alpha.f = 0.2),
  alpha_min = 0.2,
  lwd = 2,
  lty = 1,
  ...
)

\method{autoplot}{pithist}(
  object,
  single_graph = FALSE,
  style = c("histogram", "lines"),
  confint = TRUE,
  ref = TRUE,
  xlim = c(0, 1),
  ylim = c(0, NA),
  xlab = NULL,
  ylab = NULL,
  main = NULL,
  colour = "black",
  fill = "darkgray",
  border = "black",
  alpha_min = 0.2,
  size = NULL,
  linetype = 1,
  legend = FALSE,
  ...
)
}
\arguments{
\item{single_graph}{logical. Should all computed extended reliability
diagrams be plotted in a single graph?}

\item{style}{character specifying the syle of rootogram (see below). FIXME:
Description}

\item{confint}{logical. Should confident intervals be drawn?}

\item{ref, col, fill, border, alpha_min, lwd, lty, axes, box}{additional graphical
parameters for base plots, whereby \code{x} is a object of class \code{pithist}.}

\item{xlim, ylim}{graphical parameters. These may pertain either to the whole
plot or just the histogram or just the fitted line.}

\item{xlab, ylab, main}{graphical parameters.}

\item{\dots}{further graphical parameters.}

\item{object, x}{an object of class \code{pithist}.}

\item{colour, size, linetype, legend}{graphical parameters passed for 
\code{ggplot2} style plots, whereby \code{object} is a object of class \code{pithist}.}
}
\description{
PIT histograms graphically compare empirical probabilities from fitted
models with a uniform distribution.
}
\details{
PIT histograms graphically the probability integral transform (PIT), i.e.,
observed probabilities from fitted probability models, with a uniform
distribution. It leverages the \code{\link{procast}} generic and then
essentially draws a \code{\link[graphics]{hist}}.

In case of discrete distributions the PIT is either drawn randomly from the
corresponding interval or distributed proportionally in the histogram
(FIXME: not yet implemented).
}
\examples{

require("crch")
m1 <- lm(dist ~ speed, data = cars)
m2 <- crch(dist ~ speed | speed, data = cars)
m3 <- crch(dist ~ speed | speed, left = 30, data = cars)

pit1 <- pithist(m1)
pit2 <- pithist(m2, plot = FALSE)
pit3 <- pithist(m3, plot = FALSE)

plot(pit1, confint = "red", ref = "blue", fill = "lightblue")

plot(c(pit1, pit2, pit3), col = c(1, 2, 3), style = "lines")

plot(c(pit1, pit2), col = c(1, 2), single_graph = TRUE)
lines(pit3, col = 3)

}
\references{
Czado C, Gneiting T, Held L (2009). \dQuote{Predictive Model
Assessment for Count Data.} \emph{Biometrics}, \bold{65}(4), 1254--1261.

Agresti A, Coull A B (1998). \dQuote{Approximate is Better than ``Exact''
for Interval Estimation of Binomial Proportions.} \emph{The American
Statistician}, \bold{52}(2), 119--126.
}
\seealso{
\code{\link{procast}}, \code{\link[graphics]{hist}}
}
\keyword{hplot}
