% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqrplot.R
\name{qqrplot}
\alias{qqrplot}
\alias{qqrplot.default}
\alias{c.qqrplot}
\title{Q-Q Plots for Quantile Residuals}
\usage{
qqrplot(object, ...)

\method{qqrplot}{default}(
  object,
  newdata = NULL,
  plot = TRUE,
  class = NULL,
  detrend = FALSE,
  scale = c("normal", "uniform"),
  nsim = 1L,
  delta = NULL,
  simint = TRUE,
  simint_level = 0.95,
  simint_nrep = 250,
  confint = TRUE,
  ref = TRUE,
  xlab = "Theoretical quantiles",
  ylab = if (!detrend) "Quantile residuals" else "Deviation",
  main = NULL,
  ...
)
}
\arguments{
\item{object}{an object from which probability integral transforms can be
extracted using the generic function \code{\link{procast}}.}

\item{newdata}{optionally, a data frame in which to look for variables with
which to predict. If omitted, the original observations are used.}

\item{plot}{Should the \code{plot} or \code{autoplot} method be called to
draw the computed Q-Q plot? Either set \code{plot} expicitly to \code{"base"} vs.
\code{"ggplot2"} to choose the type of plot, or for a logical \code{plot} argument
it's chosen conditional if the package \code{ggplot2} is loaded.}

\item{class}{Should the invisible return value be either a \code{data.frame}
or a \code{tibble}. Either set \code{class} expicitly to \code{"data.frame"} vs.
\code{"tibble"}, or for \code{NULL} it's chosen automatically conditional if the package
\code{tibble} is loaded.}

\item{detrend}{logical. Should the qqrplot be detrended, i.e, plotted as a `wormplot()`?}

\item{scale}{On which scale should the quantile residuals be shown: on the probability scale 
(\code{"uniform"}) or on the normal scale (\code{"normal"}).}

\item{nsim, delta}{arguments passed to \code{qresiduals}.}

\item{simint}{logical. In case of discrete distributions, should the simulation
(confidence) interval due to the randomization be visualized?}

\item{simint_level}{numeric. The confidence level required for calculating the simulation
(confidence) interval due to the randomization.}

\item{simint_nrep}{numeric. The repetition number of simulated quantiles for calculating the
simulation (confidence) interval due to the randomization.}

\item{confint}{logical or character string describing the style for plotting `c("polygon", "line")`.
If not set to `FALSE`, the pointwise confidence interval of the (randomized)
quantile residuals are visualized.}

\item{ref}{logical. Should a reference line be plotted?}

\item{xlab, ylab, main, \dots}{graphical parameters passed to
\code{\link{plot.qqrplot}} or \code{\link{autoplot.qqrplot}}.}
}
\value{
An object of class \code{"qqrplot"} inheriting from
\code{"data.frame"} or \code{"tibble"} conditional on the argument \code{class}
with the following variables: \item{x}{theoretical quantiles,}
\item{y}{deviations between theoretical and empirical quantiles.} In case of
randomized residuals, \code{nsim} different \code{x} and \code{y} values, and
lower and upper confidence interval bounds (\code{simint_expected}, \code{simint_observed_lwr},
\code{simint_observed_upr}) can optionally be returned.  Additionally,
\code{xlab}, \code{ylab}, \code{main}, and \code{simint_level}, as well as the
the (\code{scale}) and wether a \code{detrended} Q-Q residuals plot
was computed are stored as attributes.
}
\description{
Visualize goodness of fit of regression models by Q-Q plots using quantile
residuals. If \code{plot = TRUE}, the resulting object of class
\code{"qqrplot"} is plotted by \code{\link{plot.qqrplot}} or
\code{\link{autoplot.qqrplot}} before it is returned, depending on whether the
package \code{ggplot2} is loaded.
}
\details{
Q-Q residuals plots draw quantile residuals (by default on the standard normal
scale) against theoretical quantiles from the same distribution.
Alternatively, quantile residuals can also be compared on the uniform scale
(\code{scale = "uniform"}) using no transformation.  For computation,
\code{\link{qqrplot}} leverages the function \code{\link{qresiduals}} employing
the \code{\link{procast}} generic.

Additional options are offered for models with discrete responses where
randomization of quantiles is needed.

In addition to the \code{plot} and \code{\link[ggplot2]{autoplot}} method for
qqrplot objects, it is also possible to combine two (or more) Q-Q residuals plots by
\code{c}/\code{rbind}, which creates a set of Q-Q residuals plots that can then be
plotted in one go.
}
\examples{

## speed and stopping distances of cars
m1_lm <- lm(dist ~ speed, data = cars)

## compute and plot qqrplot
qqrplot(m1_lm)

#-------------------------------------------------------------------------------
## determinants for male satellites to nesting horseshoe crabs
data("CrabSatellites", package = "countreg")

## linear poisson model
m1_pois <- glm(satellites ~ width + color, data = CrabSatellites, family = poisson)
m2_pois <- glm(satellites ~ color, data = CrabSatellites, family = poisson)

## compute and plot qqrplot as base graphic
q1 <- qqrplot(m1_pois, plot = FALSE)
q2 <- qqrplot(m2_pois, plot = FALSE)

## plot combined qqrplot as "ggplot2" graphic
ggplot2::autoplot(c(q1, q2), single_graph = TRUE, col = c(1, 2), fill = c(1, 2))

## Use different `scale`s with confidence intervals
qqrplot(m1_pois, scale = "uniform")
qqrplot(m1_pois, scale = "normal")
qqrplot(m1_pois, detrend = TRUE, scale = "uniform", confint = "line")
qqrplot(m1_pois, detrend = TRUE, scale = "normal", confint = "line")

}
\references{
Dunn KP, Smyth GK (1996). \dQuote{Randomized Quantile
Residuals.} \emph{Journal of Computational and Graphical Statistics},
\bold{5}, 1--10. \doi{10.2307/1390802}
}
\seealso{
\code{\link{plot.qqrplot}}, \code{\link{wormplot}},
\code{\link{qresiduals}}, \code{\link[stats]{qqnorm}}
}
\keyword{hplot}
