% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qresiduals.R
\name{qresiduals}
\alias{qresiduals}
\alias{qresiduals.default}
\title{(Randomized) Quantile Residuals}
\usage{
qresiduals(object, ...)

\method{qresiduals}{default}(object, scale = "normal", ...)
}
\arguments{
\item{object}{an object. For the \code{default} method this needs to be
either a specification of probabilities (vector or 2-dimensional matrix of
probabilities) or an object from which the these can be obtained with
\code{\link{procast}}.}

\item{\dots}{further parameters passed to  \code{\link{pitresiduals}}.}

\item{scale}{On which scale should the PIT residuals be shown; on the probability scale 
(\code{"uniform"}) or on the normal scale (\code{"normal"}).
Here, for (randomized) quantile residuals, the
quantiles of the standard normal distribution are computed per default.}
}
\value{
A vector or matrix of quantile residuals.
}
\description{
Generic function and methods for computing (randomized) quantile residuals.
}
\details{
(Randomized) quantile residuals are simply the theoretical standard normal
quantiles evaluated at the PIT residuals as suggested by Dunn and Smyth (1996).
For regression models with a continuous response distribution these are exact;
for discrete distributions, PIT residuals are drawn from the range of
probabilities corresponding to each observation, hence quantile residuals must
be random as well.

The default \code{qresiduals} method calls \code{\link{pitresiduals}} with
\code{scale} equal \code{"normal"}, as employed in normal Q-Q
plots (\code{\link{qqrplot}}).
}
\note{
Note that there is also a \code{\link[statmod]{qresiduals}} function
in the \pkg{statmod} package that is not generic and always returns a single
random quantile residual.
}
\examples{

## linear regression models (homoscedastic Gaussian response)
m <- lm(dist ~ speed, data = cars)
qresiduals(m)

}
\references{
Dunn KP, Smyth GK (1996). \dQuote{Randomized Quantile
Residuals.} \emph{Journal of Computational and Graphical Statistics},
\bold{5}(3), 236--244. \doi{10.2307/1390802}
}
\seealso{
\code{\link{pitresiduals}}, \code{\link[stats]{qnorm}}, \code{\link{qqrplot}}
}
\keyword{regression}
