\name{reliagram}
\alias{reliagram}
\alias{reliagram.default}

\alias{plot.reliagram}
\alias{autoplot.reliagram}

\title{Reliagram (Extended Reliability Diagram)}

\description{
  Reliagram (extended reliability diagram) for assessing the reliability
  of a fitted probabilistic distributional forecast.
}

\details{
TODO.
}

\keyword{hplot}
