% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rootogram.R
\name{rootogram}
\alias{rootogram}
\alias{rootogram.default}
\alias{c.rootogram}
\alias{rbind.rootogram}
\title{Rootograms for Assessing Goodness of Fit of Probability Models}
\usage{
rootogram(object, ...)

\method{rootogram}{default}(
  object,
  newdata = NULL,
  plot = TRUE,
  class = NULL,
  breaks = NULL,
  width = NULL,
  style = c("hanging", "standing", "suspended"),
  scale = c("sqrt", "raw"),
  expected = TRUE,
  confint = TRUE,
  ref = TRUE,
  xlab = NULL,
  ylab = NULL,
  main = NULL,
  ...
)
}
\arguments{
\item{object}{an object from which an rootogram can be extracted with
\code{\link{procast}}.}

\item{\dots}{further graphical parameters passed to the plotting function.}

\item{newdata}{an optional data frame in which to look for variables with
which to predict. If omitted, the original observations are used.}

\item{plot}{logical or character. Should the \code{plot} or \code{autoplot} method be called to
draw the computed extended reliability diagram? Logical
\code{FALSE} will suppress plotting, \code{TRUE} (default) will choose the
type of plot conditional if the package \code{ggplot2} is loaded.
Alternatively \code{"base"} or \code{"ggplot2"} can be specified to
explicitly choose the type of plot.}

\item{class}{should the invisible return value be either a \code{data.frame}
or a \code{tbl_df}. Can be set to \code{"data.frame"} or \code{"tibble"} to
explicitly specify the return class, or to \code{NULL} (default) in which
case the return class is conditional on whether the package \code{"tibble"}
is loaded.}

\item{breaks}{\code{NULL} (default) or numeric to manually specify the breaks for
the rootogram intervals. A single numeric (larger \code{0}) specifies the number of breaks
to be automatically chosen, multiple numeric values are interpreted as manually specified breaks.}

\item{width}{\code{NULL} (default) or single positive numeric. Width of the histogram bars.
Will be ignored for non-discrete distributions.}

\item{style}{character specifying the syle of rootogram (see 'Details').}

\item{scale}{character specifying whether \code{"raw"} frequencies or their square
roots (\code{"sqrt"}; default) should be drawn.}

\item{expected}{logical or character. Should the expected (fitted) frequencies be plotted?
Can be set to \code{"both"} (same as \code{TRUE}; default), \code{"line"}, \code{"point"},
or \code{FALSE} which will suppress plotting.}

\item{confint}{logical, defaults to \code{TRUE}. Should confident intervals be drawn?}

\item{ref}{logical, defaults to \code{TRUE}. Should a reference line be plotted?}

\item{xlab, ylab, main}{graphical parameters forwarded to
\code{\link{plot.rootogram}} or \code{\link{autoplot.rootogram}}.}
}
\value{
An object of class \code{"rootogram"} inheriting from
\code{"data.frame"} or \code{"tibble"} conditional on the argument \code{class}
with the following variables:
\item{observed}{observed frequencies,}
\item{expected}{expected (fitted) frequencies,}
\item{mid}{histogram interval midpoints on the x-axis,}
\item{width}{widths of the histogram bars,}
\item{confint_lwr, confint_upr}{lower and upper confidence interval bound.} 

Additionally, \code{style}, \code{scale}, \code{expected}, \code{confint},
\code{ref}, \code{xlab}, \code{ylab}, amd \code{main} are stored as attributes.
}
\description{
Rootograms graphically compare (square roots) of empirical frequencies with
expected (fitted) frequencies from a probabilistic model. If \code{plot = TRUE}, the
resulting object of class \code{"rootogram"} is plotted by
\code{\link{plot.rootogram}} or \code{\link{autoplot.rootogram}} before it is
returned, depending on whether the package \code{ggplot2} is loaded.
}
\details{
Rootograms graphically compare frequencies of empirical distributions and
expected (fitted) probability models. For the observed distribution the histogram is
drawn on a square root scale (hence the name) and superimposed with a line
for the expected frequencies. The histogram can be \code{"hanging"} from the
expected curve (default), \code{"standing"} on the (like bars in barplot),
or drawn as a \code{"suspended"} histogram of deviations.

The function \code{\link{rootogram}} leverages the \code{\link{procast}}
generic in order to compute all necessary coordinates based on observed and
expected (fitted) frequencies.

In addition to the \code{plot} and \code{\link[ggplot2]{autoplot}} method for
rootogram objects, it is also possible to combine two (or more) rootograms by
\code{c}/\code{rbind}, which creates a set of rootograms that can then be
plotted in one go.
}
\note{
Note that there is also a \code{\link[vcd]{rootogram}} function in the
\pkg{vcd} package that is similar to the \code{numeric} method provided
here. However, it is much more limited in scope, hence a function has been
created here.
}
\examples{
## plots and output

## number of deaths by horsekicks in Prussian army (Von Bortkiewicz 1898)
deaths <- rep(0:4, c(109, 65, 22, 3, 1))

## fit glm model
m1_pois <- glm(deaths ~ 1, family = poisson)
rootogram(m1_pois)

## inspect output (without plotting)
r1 <- rootogram(m1_pois, plot = FALSE)
r1

## combine plots
plot(c(r1, r1), col = c(1, 2), expected_col = c(1, 2))


#-------------------------------------------------------------------------------
## different styles

## artificial data from negative binomial (mu = 3, theta = 2)
## and Poisson (mu = 3) distribution
set.seed(1090)
y <- rnbinom(100, mu = 3, size = 2)
x <- rpois(100, lambda = 3)

## glm method: fitted values via glm()
m2_pois <- glm(y ~ x, family = poisson)

## correctly specified Poisson model fit
par(mfrow = c(1, 3))
r1 <- rootogram(m2_pois, style = "standing", ylim = c(-2.2, 4.8), main = "Standing")
r2 <- rootogram(m2_pois, style = "hanging", ylim = c(-2.2, 4.8), main = "Hanging")
r3 <- rootogram(m2_pois, style = "suspended", ylim = c(-2.2, 4.8), main = "Suspended")
par(mfrow = c(1, 1))

#-------------------------------------------------------------------------------
## linear regression with normal/Gaussian response: anorexia data

data("anorexia", package = "MASS")

m3_gauss <- glm(Postwt ~ Prewt + Treat + offset(Prewt), family = gaussian, data = anorexia)

## plot rootogram as "ggplot2" graphic
rootogram(m3_gauss, plot = "ggplot2")
}
\references{
Friendly M (2000), \emph{Visualizing Categorical Data}. SAS
Institute, Cary, ISBN 1580256600.

Kleiber C, Zeileis A (2016).  \dQuote{Visualizing Count Data Regressions
Using Rootograms.} \emph{The American Statistician}, \bold{70}(3), 296--303.
\doi{10.1080/00031305.2016.1173590}.

Tukey JW (1977). \emph{Exploratory Data Analysis}. Addison-Wesley, Reading,
ISBN 0201076160.
}
\seealso{
\code{\link{plot.rootogram}}, \code{\link{procast}}
}
\keyword{hplot}
