% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pithist.R
\docType{data}
\name{stat_pithist}
\alias{stat_pithist}
\alias{StatPithist}
\alias{geom_pithist}
\alias{GeomPithist}
\alias{stat_pithist_ref}
\alias{StatPithistRef}
\alias{geom_pithist_ref}
\alias{GeomPithistRef}
\alias{stat_pithist_confint}
\alias{StatPithistConfint}
\alias{geom_pithist_confint}
\alias{GeomPithistConfint}
\alias{stat_pithist_simint}
\alias{StatPithistSimint}
\alias{geom_pithist_simint}
\alias{GeomPithistSimint}
\title{\code{geom_*} and \code{stat_*} for Producing PIT Histograms with `ggplot2`}
\usage{
stat_pithist(
  mapping = NULL,
  data = NULL,
  geom = "pithist",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  freq = FALSE,
  style = c("bar", "line"),
  ...
)

geom_pithist(
  mapping = NULL,
  data = NULL,
  stat = "pithist",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  freq = FALSE,
  style = c("bar", "line"),
  ...
)

stat_pithist_ref(
  mapping = NULL,
  data = NULL,
  geom = "pithist_ref",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  trafo = NULL,
  freq = FALSE,
  ...
)

geom_pithist_ref(
  mapping = NULL,
  data = NULL,
  stat = "pithist_ref",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  trafo = NULL,
  freq = FALSE,
  ...
)

stat_pithist_confint(
  mapping = NULL,
  data = NULL,
  geom = "pithist_confint",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  trafo = NULL,
  level = 0.95,
  type = "approximation",
  freq = FALSE,
  style = c("polygon", "line"),
  ...
)

geom_pithist_confint(
  mapping = NULL,
  data = NULL,
  stat = "pithist_confint",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  trafo = NULL,
  level = 0.95,
  type = "approximation",
  freq = FALSE,
  style = c("polygon", "line"),
  ...
)

stat_pithist_simint(
  mapping = NULL,
  data = NULL,
  geom = "pithist_simint",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  freq = FALSE,
  ...
)

geom_pithist_simint(
  mapping = NULL,
  data = NULL,
  stat = "pithist_simint",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  freq = FALSE,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}} or
\code{\link[ggplot2:aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{geom}{The geometric object to use display the data}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}

\item{freq}{Fix description.}

\item{style}{Fix description.}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{stat}{The statistical transformation to use on the data for this
layer, as a string.}

\item{trafo}{Fix description.}

\item{level}{Fix description.}

\item{type}{Fix description.}
}
\description{
Various \code{geom_*} and \code{stat_*} used within
\code{\link[ggplot2]{autoplot}} for producing PIT histograms.
}
\examples{
if (require("ggplot2")) {
  ## Fit model
  data("CrabSatellites", package = "countreg")
  m1_pois <- glm(satellites ~ width + color, data = CrabSatellites, family = poisson)
  m2_pois <- glm(satellites ~ color, data = CrabSatellites, family = poisson)
  
  ## Compute pithist
  p1 <- pithist(m1_pois, type = "random", plot = FALSE)
  p2 <- pithist(m2_pois, type = "random", plot = FALSE)
  
  d <- c(p1, p2) 
  
  ## Create factor
  main <- attr(d, "main")
  main <- make.names(main, unique = TRUE)
  d$group <- factor(d$group, labels = main)
  
  ## Plot bar style PIT histogram
  gg1 <- ggplot(data = d) + 
    geom_pithist(aes(x = mids, y = counts, width = width, group = group), freq = TRUE) + 
    geom_pithist_simint(aes(x = mids, ymin = simint_lwr, ymax = simint_upr), freq = TRUE) + 
    geom_pithist_confint(aes(x = mids, y = counts, width = width), style = "line", freq = TRUE) + 
    geom_pithist_ref(aes(x = mids, y = counts, width = width), freq = TRUE) + 
    facet_grid(group~.) + xlab("PIT") + ylab("Frequency")
  gg1

  gg2 <- ggplot(data = d) + 
    geom_pithist(aes(x = mids, y = counts, width = width, group = group), freq = FALSE) + 
    geom_pithist_simint(aes(x = mids, ymin = simint_lwr, ymax = simint_upr, y = counts, 
      width = width), freq = FALSE) + 
    geom_pithist_confint(aes(x = mids, y = counts, width = width), style = "line", freq = FALSE) + 
    geom_pithist_ref(aes(x = mids, y = counts, width = width), freq = FALSE) + 
    facet_grid(group~.) + xlab("PIT") + ylab("Density")
  gg2

  ## Plot line style PIT histogram
  gg3 <- ggplot(data = d) + 
    geom_pithist(aes(x = mids, y = counts, width = width, group = group), style = "line") + 
    geom_pithist_confint(aes(x = mids, y = counts, width = width), style = "polygon") + 
    facet_grid(group~.) + xlab("PIT") + ylab("Density")
  gg3
}
}
\keyword{datasets}
