% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rootogram.R
\docType{data}
\name{stat_rootogram}
\alias{stat_rootogram}
\alias{StatRootogram}
\alias{geom_rootogram}
\alias{GeomRootogram}
\alias{stat_rootogram_expected}
\alias{StatRootogramExpected}
\alias{geom_rootogram_expected}
\alias{GeomRootogramExpected}
\alias{geom_rootogram_ref}
\alias{GeomRootogramRef}
\title{\code{geom_*} and \code{stat_*} for Producing PIT Histograms with `ggplot2`}
\format{
An object of class \code{GeomRootogramExpected} (inherits from \code{GeomPath}, \code{Geom}, \code{ggproto}, \code{gg}) of length 3.
}
\usage{
stat_rootogram(
  mapping = NULL,
  data = NULL,
  geom = "rootogram",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  scale = c("sqrt", "raw"),
  style = c("hanging", "standing", "suspended"),
  ...
)

geom_rootogram(
  mapping = NULL,
  data = NULL,
  stat = "rootogram",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  scale = c("sqrt", "raw"),
  style = c("hanging", "standing", "suspended"),
  ...
)

stat_rootogram_expected(
  mapping = NULL,
  data = NULL,
  geom = "rootogram_expected",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  scale = c("sqrt", "raw"),
  ...
)

geom_rootogram_expected(
  mapping = NULL,
  data = NULL,
  stat = "rootogram_expected",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  scale = c("sqrt", "raw"),
  linestyle = c("both", "line", "point"),
  ...
)

GeomRootogramExpected

geom_rootogram_ref(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}} or
\code{\link[ggplot2:aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{geom}{The geometric object to use display the data}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}

\item{scale}{character specifying whether values should be transformed to the square root scale (not checking for original scale, so maybe applied again).}

\item{style}{character specifying the syle of rootogram (see below).}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{stat}{The statistical transformation to use on the data for this
layer, as a string.}

\item{linestyle}{Character string defining one of `"both"`, `"line"` or `"point"`.}
}
\description{
Various \code{geom_*} and \code{stat_*} used within
\code{\link[ggplot2]{autoplot}} for producing PIT histograms.
}
\examples{
if (require("ggplot2")) {
  ## Fit model
  data("CrabSatellites", package = "countreg")
  m1_pois <- glm(satellites ~ width + color, data = CrabSatellites, family = poisson)
  m2_pois <- glm(satellites ~ color, data = CrabSatellites, family = poisson)

  ## Compute rootogram (on raw scale)
  p1 <- rootogram(m1_pois, scale = "raw", plot = FALSE)
  p2 <- rootogram(m2_pois, scale = "raw", plot = FALSE)

  d <- c(p1, p2)

  ## Get label names
  main <- attr(d, "main")
  main <- make.names(main, unique = TRUE)
  d$group <- factor(d$group, labels = main)

  ## Plot rootograms w/ on default "sqrt" scale
  gg1 <- ggplot(data = d) +
    geom_rootogram(aes(
      observed = observed, expected = expected, mid = mid,
      width = width, group = group
    )) +
    geom_rootogram_expected(aes(expected = expected, mid = mid)) +
    geom_rootogram_ref() +
    facet_grid(group ~ .) + 
    xlab("satellites") +
    ylab("sqrt(Frequency)")
  gg1
}
}
\keyword{datasets}
