% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NumericNormal.R
\name{NumericNormal}
\alias{NumericNormal}
\alias{pdf.NumericNormal}
\alias{cdf.NumericNormal}
\alias{quantile.NumericNormal}
\alias{mean.NumericNormal}
\alias{variance.NumericNormal}
\alias{skewness.NumericNormal}
\alias{kurtosis.NumericNormal}
\title{Normal Distribution for Testing Numeric Moment Calculations}
\usage{
NumericNormal(mu, var)

\method{pdf}{NumericNormal}(d, x, ...)

\method{cdf}{NumericNormal}(d, x, ..., n = 101L)

\method{quantile}{NumericNormal}(x, probs, ..., n = 101L)

\method{mean}{NumericNormal}(x, ..., n = 101L)

\method{variance}{NumericNormal}(x, ..., n = 101L)

\method{skewness}{NumericNormal}(x, ..., n = 101L)

\method{kurtosis}{NumericNormal}(x, ..., n = 101L)
}
\arguments{
\item{mu}{numeric vector with means.}

\item{var}{numeric vector with variances.}

\item{d}{object of class `"NumericNormal"`.}

\item{x}{object of class `c("NumericNormal", "distribution")`.}

\item{\dots}{ignored.}

\item{n}{integer, number of grid points used for numerical integration.}

\item{probs}{numeric vector, probabilities.}
}
\value{
Object of class `c("NumericNormal", "distribution")`.
}
\description{
Normal Distribution for Testing Numeric Moment Calculations
}
\examples{
\dontrun{
n <- Normal(5, sqrt(4))
r <- NumericNormal(5, 4)

## ---------------- CDF ----------------
## 'NumericNormal' only knows the (exact) pdf of a Gaussian distribution
## and approximates the CDF via numeric integration (default grid size n = 101L).
cdf(n, 3.5)
cdf(r, 3.5)

## Visual example
xx <- seq(-5, 15, by = 0.1)
plot(xx, cdf(n, xx), main = "Testing numeric PDF->CDF")
lines(xx, cdf(r, xx), col = 2, lwd = 3)
lines(xx, cdf(r, xx, n = 5L), col = 4, lwd = 3)
legend("topleft", legend = c("Normal", "NumericNormal n = 101", "NumericNormal n = 5"),
       pch = c(1, NA, NA), lty = c(NA, 1, 1), lwd = c(NA, 3, 3), col = c(1, 2, 4))

## -------------- Quantile -------------
## 'NumericNormal' only knows the PDF from which it calculates the CDF
## to retrieve the quantiles.
quantile(n, 0.3)
quantile(r, 0.3)

qq <- c(0.0001, seq(0.01, 0.99, by = 0.005), 0.9999)
plot(quantile(n, qq), qq, main = "Testing numeric PDF->Quantile")
lines(quantile(r, qq), qq, col = 2, lwd = 3)
lines(quantile(r, qq, n = 5L), qq, col = 4, lwd = 3)
legend("topleft",
       legend = c("Normal", "NumericNormal n = 101", "NumericNormal n = 5"),
       pch = c(1, NA, NA), lty = c(NA, 1, 1), lwd = c(NA, 3, 3), col = c(1, 2, 4))

## Central moments
c(mean = mean(r), variance = variance(r),
  skewness = skewness(r), kurtosis = kurtosis(r))
}

}
