% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pithist.R
\name{pithist}
\alias{pithist}
\alias{pithist.default}
\alias{c.pithist}
\title{PIT Histograms for Assessing Goodness of Fit of Probability Models}
\usage{
pithist(object, ...)

\method{pithist}{default}(
  object,
  newdata = NULL,
  plot = TRUE,
  class = NULL,
  style = c("histogram", "lines"),
  type = c("random", "proportional"),
  nsim = 1L,
  delta = NULL,
  freq = FALSE,
  breaks = NULL,
  confint = TRUE,
  confint_level = 0.95,
  confint_type = c("exact", "approximation"),
  single_graph = FALSE,
  xlim = c(0, 1),
  ylim = c(0, NA),
  xlab = "PIT",
  ylab = if (freq) "Frequency" else "Density",
  main = NULL,
  ...
)
}
\arguments{
\item{object}{an object from which probability integral transforms can be
extracted with \code{\link{procast}}.}

\item{\dots}{further graphical parameters.}

\item{newdata}{optionally, a data frame in which to look for variables with
which to predict. If omitted, the original observations are used.}

\item{plot}{Should the \code{plot} or \code{autoplot} method be called to
draw the computed extended reliability diagram? Either set \code{plot}
expicitly to "base" vs. "ggplot2" to choose the type of plot, or for a
logical \code{plot} argument it's chosen conditional if the package
\code{ggplot2} is loaded.}

\item{class}{Should the invisible return value be either a \code{data.frame}
or a \code{tibble}. Either set \code{class} expicitly to "data.frame" vs.
"tibble", or for NULL it's chosen automatically conditional if the package
\code{tibble} is loaded.}

\item{style}{character specifying the syle of rootogram (see below). FIXME:
Description}

\item{type}{character. In case of discrete distributions should the PITs be
drawn randomly from the corresponding interval or distributed
proportionally?}

\item{nsim}{integer. If \code{type} is \code{"random"} how many simulated
PITs should be drawn?}

\item{delta}{numeric. The minimal difference to compute the range of
proabilities corresponding to each observation according to get (randomized)
quantile residuals.  For \code{NULL}, the minimal observed difference in the
resonse divided by \code{5e-6} is used.}

\item{freq}{logical. If \code{TRUE}, the PIT histogram is represented by
frequencies, the \code{counts} component of the result; if \code{FALSE},
probability densities, component \code{density}, are plotted (so that the
histogram has a total area of one).}

\item{breaks}{numeric. Breaks for the histogram intervals.}

\item{confint}{logical. Should confident intervals be drawn?}

\item{confint_level}{numeric. The confidence level required.}

\item{confint_type}{character. Which type of confidence interval.  According
to Agresti and Coull (1998) for interval estimation of binomial proportions
an approximation can be better than exact.}

\item{single_graph}{logical. Should all computed extended reliability
diagrams be plotted in a single graph?}

\item{xlim, ylim}{graphical parameters. These may pertain either to the whole
plot or just the histogram or just the fitted line.}

\item{xlab, ylab, main}{graphical parameters.}
}
\description{
PIT histograms graphically compare empirical probabilities from fitted
models with a uniform distribution.
}
\details{
PIT histograms graphically evaluate the probability integral transform (PIT),
i.e., the value that the predictive CDF attains at the observation, with a
uniform distribution. For a well calibrated model fit, the observation will be
drawn from the predictive distribution and the PIT will have a standard uniform
distribution. \code{\link{pithist}} leverages the function
\code{\link{qresiduals}} employing the \code{\link{procast}} generic and then
essentially draws a \code{\link[graphics]{hist}}. 

In case of discrete distributions the PIT can be either drawn randomly from the
corresponding interval or distributed proportionally in the histogram, whereby
the latter is not fully supported yet.
}
\examples{

require("crch")
m1 <- lm(dist ~ speed, data = cars)
m2 <- crch(dist ~ speed | speed, data = cars)
m3 <- crch(dist ~ speed | speed, left = 30, data = cars)

pit1 <- pithist(m1)
pit2 <- pithist(m2, plot = FALSE)
pit3 <- pithist(m3, plot = FALSE)

plot(pit1, confint = "red", ref = "blue", fill = "lightblue")

plot(c(pit1, pit2), col = c(1, 2), single_graph = TRUE, style = "lines")
lines(pit3, col = 3)

}
\references{
Czado C, Gneiting T, Held L (2009). \dQuote{Predictive Model
Assessment for Count Data.} \emph{Biometrics}, \bold{65}(4), 1254--1261. 
\doi{10.2307/2981683}.
 
Dawid AP (1984). \dQuote{Present Position and Potential Developments: Some
Personal Views: Statistical Theory: The Prequential Approach}, \emph{Journal of
the Royal Statistical Society: Series A (General)}, \bold{147}(2), 278--292,
\doi{10.2307/2981683}

Diebold FX, Gunther TA, Tay AS (1998). \dQuote{Evaluating Density Forecasts
with Applications to Financial Risk Management}. \emph{International Economic
Review}, \bold{39}(4), 863--883. \doi{10.2307/2527342}.

Tilmann G, Balabdaoui F, Raftery AE (2007). \dQuote{Probabilistic Forecasts,
Calibration and Sharpness}.  \emph{Journal of the Royal Statistical Society:
Series B (Methodological)}. \bold{69}(2), 243--268,
\doi{10.1111/j.1467-9868.2007.00587.x}.
}
\seealso{
\code{\link{procast}}, \code{\link[graphics]{hist}}
}
\keyword{hplot}
