% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pithist.R
\name{pithist}
\alias{pithist}
\alias{pithist.default}
\alias{c.pithist}
\alias{rbind.pithist}
\title{PIT Histograms for Assessing Goodness of Fit of Probability Models}
\usage{
pithist(object, ...)

\method{pithist}{default}(
  object,
  newdata = NULL,
  plot = TRUE,
  class = NULL,
  trafo = NULL,
  breaks = NULL,
  type = c("expected", "random"),
  nsim = 1L,
  delta = NULL,
  simint = NULL,
  simint_level = 0.95,
  simint_nrep = 250,
  style = c("bar", "line"),
  freq = FALSE,
  expected = TRUE,
  confint = TRUE,
  xlab = "PIT",
  ylab = if (freq) "Frequency" else "Density",
  main = NULL,
  ...
)
}
\arguments{
\item{object}{an object from which probability integral transforms can be
extracted using the generic function \code{\link{procast}}.}

\item{\dots}{further graphical parameters.}

\item{newdata}{optionally, a data frame in which to look for variables with
which to predict. If omitted, the original observations are used.}

\item{plot}{Should the \code{plot} or \code{autoplot} method be called to
draw the computed extended reliability diagram? Either set \code{plot}
expicitly to \code{"base"} vs. \code{"ggplot2"} to choose the type of plot, or for a
logical \code{plot} argument it's chosen conditional if the package
\code{ggplot2} is loaded.}

\item{class}{Should the invisible return value be either a \code{data.frame}
or a \code{tibble}. Either set \code{class} expicitly to \code{"data.frame"} vs.
\code{"tibble"}, or for \code{NULL} it's chosen automatically conditional if the package
\code{tibble} is loaded.}

\item{trafo}{function for tranforming residuals from probability scale to a
different distribution scale.}

\item{breaks}{numeric. Breaks for the histogram intervals.}

\item{type}{character. In case of discrete distributions, should an expected
(non-normal) PIT histogram be computed according to Czado et al. (2009)
(\code{"expected"}) or should the PIT be drawn randomly from the corresponding
interval (\code{"random"})?}

\item{nsim}{integer. If \code{type} is \code{"random"} how many simulated
PITs should be drawn?}

\item{delta}{numeric. The minimal difference to compute the range of
proabilities corresponding to each observation according to get (randomized)
quantile residuals. For \code{NULL}, the minimal observed difference in the
resonse divided by \code{5e-6} is used.}

\item{simint}{logical. In case of discrete distributions, should the simulation
(confidence) interval due to the randomization be visualized?}

\item{simint_level}{numeric. The confidence level required for calculating the simulation
(confidence) interval due to the randomization.}

\item{simint_nrep}{numeric. The repetition number of simulated quantiles for calculating the
simulation (confidence) interval due to the randomization.}

\item{style}{character specifying the style of pithist. For \code{style = "bar"}
a traditional PIT hisogram is drawn, for \code{style = "line"} solely the upper border
line is plotted. For \code{single_graph = TRUE}, always line-style PIT histograms are
drawn.}

\item{freq}{logical. If \code{TRUE}, the PIT histogram is represented by
frequencies, the \code{counts} component of the result; if \code{FALSE},
probability densities, component \code{density}, are plotted (so that the
histogram has a total area of one).}

\item{expected}{logical. Should the expected values be plotted as reference?}

\item{confint}{logical. Should confident intervals be drawn?}

\item{xlab, ylab, main}{graphical parameters passed to
\code{\link{plot.pithist}} or \code{\link{autoplot.pithist}}.}
}
\value{
An object of class \code{"pithist"} inheriting from
\code{"data.frame"} or \code{"tibble"} conditional on the argument \code{class}
with the following variables: \item{x}{histogram
interval midpoints on the x-axis,} \item{y}{bottom coordinate of the
histogram bars,} \item{width}{widths of the histogram bars,}
\item{confint_lwr, confint_upr}{lower and upper confidence interval bound,} \item{expected}{y-coordinates of
the expected curve.} Additionally, \code{freq}, \code{xlab},
\code{ylab}, \code{main}, and \code{confint_level} are stored as attributes.
}
\description{
PIT histograms graphically compare empirical probabilities from fitted models
with a uniform distribution. If \code{plot = TRUE}, the resulting object of
class \code{"pithist"} is plotted by \code{\link{plot.pithist}} or
\code{\link{autoplot.pithist}} before it is returned, depending on whether the
package \code{ggplot2} is loaded.
}
\details{
PIT histograms graphically evaluate the probability integral transform (PIT),
i.e., the value that the predictive CDF attains at the observation, with a
uniform distribution. For a well calibrated model fit, the observation will be
drawn from the predictive distribution and the PIT will have a standard uniform
distribution. For computation, \code{\link{pithist}} leverages the function
\code{\link{qresiduals}} employing the \code{\link{procast}} generic and then
essentially draws a \code{\link[graphics]{hist}}.

In case of discrete distributions the PIT can be either drawn randomly from the
corresponding interval or distributed proportionally in the histogram, whereby
the latter is not yet supported.

In addition to the \code{plot} and \code{\link[ggplot2]{autoplot}} method for
pithist objects, it is also possible to combine two (or more) PIT histograms by
\code{c}/\code{rbind}, which creates a set of PIT histograms that can then be
plotted in one go.
}
\examples{
## speed and stopping distances of cars
m1_lm <- lm(dist ~ speed, data = cars)

## compute and plot pithist
pithist(m1_lm)

#-------------------------------------------------------------------------------
## determinants for male satellites to nesting horseshoe crabs
data("CrabSatellites", package = "countreg")

## linear poisson model
m1_pois <- glm(satellites ~ width + color, data = CrabSatellites, family = poisson)
m2_pois <- glm(satellites ~ color, data = CrabSatellites, family = poisson)

## compute and plot pithist as base graphic
p1 <- pithist(m1_pois, plot = FALSE)
p2 <- pithist(m2_pois, plot = FALSE)

## plot combined pithist as "ggplot2" graphic
ggplot2::autoplot(c(p1, p2), single_graph = TRUE, style = "line", col = c(1, 2))
}
\references{
Agresti A, Coull AB (1998). \dQuote{Approximate is Better than ``Exact''
for Interval Estimation of Binomial Proportions.} \emph{The American
Statistician}, \bold{52}(2), 119--126. \doi{10.1080/00031305.1998.10480550}

Czado C, Gneiting T, Held L (2009). \dQuote{Predictive Model
Assessment for Count Data.} \emph{Biometrics}, \bold{65}(4), 1254--1261.
\doi{10.2307/2981683}

Dawid AP (1984). \dQuote{Present Position and Potential Developments: Some
Personal Views: Statistical Theory: The Prequential Approach}, \emph{Journal of
the Royal Statistical Society: Series A (General)}, \bold{147}(2), 278--292.
\doi{10.2307/2981683}

Diebold FX, Gunther TA, Tay AS (1998). \dQuote{Evaluating Density Forecasts
with Applications to Financial Risk Management}. \emph{International Economic
Review}, \bold{39}(4), 863--883. \doi{10.2307/2527342}

Gneiting T, Balabdaoui F, Raftery AE (2007). \dQuote{Probabilistic Forecasts,
Calibration and Sharpness}.  \emph{Journal of the Royal Statistical Society:
Series B (Methodological)}. \bold{69}(2), 243--268.
\doi{10.1111/j.1467-9868.2007.00587.x}
}
\seealso{
\code{\link{plot.pithist}}, \code{\link{qresiduals}}, \code{\link{procast}}
}
\keyword{hplot}
