% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pithist.R
\name{plot.pithist}
\alias{plot.pithist}
\alias{lines.pithist}
\alias{autoplot.pithist}
\title{S3 Methods for Plotting PIT Histograms}
\usage{
\method{plot}{pithist}(
  x,
  single_graph = FALSE,
  style = NULL,
  freq = NULL,
  expected = TRUE,
  confint = NULL,
  confint_level = 0.95,
  confint_type = c("exact", "approximation"),
  simint = NULL,
  xlim = c(NA, NA),
  ylim = c(0, NA),
  xlab = NULL,
  ylab = NULL,
  main = NULL,
  axes = TRUE,
  box = TRUE,
  col = "black",
  border = "black",
  lwd = NULL,
  lty = 1,
  alpha_min = 0.2,
  expected_col = NULL,
  expected_lty = NULL,
  expected_lwd = 1.75,
  confint_col = NULL,
  confint_lty = 2,
  confint_lwd = 1.75,
  confint_alpha = NULL,
  simint_col = "black",
  simint_lty = 1,
  simint_lwd = 1.75,
  ...
)

\method{lines}{pithist}(
  x,
  freq = NULL,
  expected = FALSE,
  confint = FALSE,
  confint_level = 0.95,
  confint_type = c("exact", "approximation"),
  simint = FALSE,
  col = "black",
  lwd = 2,
  lty = 1,
  expected_col = "black",
  expected_lty = 2,
  expected_lwd = 1.75,
  confint_col = "black",
  confint_lty = 1,
  confint_lwd = 1.75,
  confint_alpha = 1,
  simint_col = "black",
  simint_lty = 1,
  simint_lwd = 1.75,
  ...
)

\method{autoplot}{pithist}(
  object,
  single_graph = FALSE,
  style = NULL,
  freq = NULL,
  expected = NULL,
  confint = NULL,
  confint_level = 0.95,
  confint_type = c("exact", "approximation"),
  simint = NULL,
  xlim = c(NA, NA),
  ylim = c(0, NA),
  xlab = NULL,
  ylab = NULL,
  main = NULL,
  legend = FALSE,
  theme = NULL,
  colour = NULL,
  fill = NULL,
  size = NULL,
  linetype = NULL,
  alpha = NULL,
  expected_colour = NULL,
  expected_size = 0.75,
  expected_linetype = NULL,
  expected_alpha = NA,
  confint_colour = NULL,
  confint_fill = NULL,
  confint_size = 0.75,
  confint_linetype = NULL,
  confint_alpha = NULL,
  simint_colour = "black",
  simint_size = 0.5,
  simint_linetype = 1,
  simint_alpha = NA,
  ...
)
}
\arguments{
\item{single_graph}{logical. Should all computed extended reliability
diagrams be plotted in a single graph? If yes, \code{style} must be set to \code{"line"}.}

\item{style}{character specifying the style of pithist. For \code{style = "bar"}
a traditional PIT hisogram is drawn, for \code{style = "line"} solely the upper border
line is plotted. For \code{single_graph = TRUE}, always line-style PIT histograms are
plotted.}

\item{freq}{logical. If \code{TRUE}, the PIT histogram is represented by
frequencies, the \code{counts} component of the result; if \code{FALSE},
probability densities, component \code{density}, are plotted (so that the
histogram has a total area of one).}

\item{expected}{logical. Should the expected values be plotted as reference?}

\item{confint}{Should confident intervals be drawn? Either logical or as 
character string defining one of `"polygon"`, `"line"` or `"none"`.}

\item{confint_level}{numeric. The confidence level required.}

\item{confint_type}{character. Which type of confidence interval should be plotted: 
`"exact"` or `"approximation"`. According
to Agresti and Coull (1998), for interval estimation of binomial proportions
an approximation can be better than exact.}

\item{simint}{logical. In case of discrete distributions, should the simulation
(confidence) interval due to the randomization be visualized?}

\item{xlim, ylim, xlab, ylab, main, axes, box}{graphical parameters.}

\item{col, border, lwd, lty, alpha_min}{graphical parameters for the main part of the base plot.}

\item{simint_col, simint_lty, simint_lwd, confint_col, confint_lty, confint_lwd, confint_alpha, expected_col, expected_lty, expected_lwd}{Further graphical parameters for the `confint` and `simint` line/polygon in the base plot.}

\item{\dots}{further graphical parameters passed to the plotting function.}

\item{object, x}{an object of class \code{\link{pithist}}.}

\item{legend}{logical. Should a legend be added in the \code{ggplot2} style
graphic?}

\item{theme}{Which `ggplot2` theme should be used. If not set, \code{\link[ggplot2]{theme_bw}} is employed.}

\item{colour, fill, size, linetype, alpha}{graphical parameters for the histogram style part in the \code{autoplot}.}

\item{simint_colour, simint_size, simint_linetype, simint_alpha, confint_colour, confint_fill, confint_size, confint_linetype, expected_colour, expected_size, expected_linetype, expected_alpha}{Further graphical parameters for the `confint` and `simint` line/polygon using \code{\link[ggplot2]{autoplot}}.}
}
\description{
Generic plotting functions for PIT histograms of the class \code{"pithist"}
computed by \code{link{pithist}}.
}
\details{
PIT histograms graphically evaluate the probability integral transform (PIT),
i.e., the value that the predictive CDF attains at the observation, with a
uniform distribution. For a well calibrated model fit, the observation will be
drawn from the predictive distribution and the PIT will have a standard uniform
distribution.

PIT histograms can be rendered as \code{ggplot2} or base R graphics by using
the generics \code{\link[ggplot2]{autoplot}} or \code{\link[graphics]{plot}}.
For a single base R graphically panel, \code{\link{lines}} adds an additional PIT histogram.
}
\examples{

## speed and stopping distances of cars
m1_lm <- lm(dist ~ speed, data = cars)

## compute and plot pithist
pithist(m1_lm)

## customize colors and style
pithist(m1_lm, expected_col = "blue", lty = 2, pch = 20, style = "line")

## add separate model
if (require("crch", quietly = TRUE)) {
  m1_crch <- crch(dist ~ speed | speed, data = cars)
  #lines(pithist(m1_crch, plot = FALSE), col = 2, lty = 2, confint_col = 2) #FIXME
}

#-------------------------------------------------------------------------------
if (require("crch")) {

  ## precipitation observations and forecasts for Innsbruck
  data("RainIbk", package = "crch")
  RainIbk <- sqrt(RainIbk)
  RainIbk$ensmean <- apply(RainIbk[, grep("^rainfc", names(RainIbk))], 1, mean)
  RainIbk$enssd <- apply(RainIbk[, grep("^rainfc", names(RainIbk))], 1, sd)
  RainIbk <- subset(RainIbk, enssd > 0)

  ## linear model w/ constant variance estimation
  m2_lm <- lm(rain ~ ensmean, data = RainIbk)

  ## logistic censored model
  m2_crch <- crch(rain ~ ensmean | log(enssd), data = RainIbk, left = 0, dist = "logistic")

  ## compute pithists
  pit2_lm <- pithist(m2_lm, plot = FALSE)
  pit2_crch <- pithist(m2_crch, plot = FALSE)

  ## plot in single graph with style "line"
  plot(c(pit2_lm, pit2_crch),
    col = c(1, 2), confint_col = c(1, 2), expected_col = 3,
    style = "line", single_graph = TRUE
  )
}

#-------------------------------------------------------------------------------
## determinants for male satellites to nesting horseshoe crabs
data("CrabSatellites", package = "countreg")

## linear poisson model
m3_pois <- glm(satellites ~ width + color, data = CrabSatellites, family = poisson)

## compute and plot pithist as "ggplot2" graphic
pithist(m3_pois, plot = "ggplot2")
}
\references{
Agresti A, Coull AB (1998). \dQuote{Approximate is Better than ``Exact''
for Interval Estimation of Binomial Proportions.} \emph{The American
Statistician}, \bold{52}(2), 119--126. \doi{10.1080/00031305.1998.10480550}

Czado C, Gneiting T, Held L (2009). \dQuote{Predictive Model
Assessment for Count Data.} \emph{Biometrics}, \bold{65}(4), 1254--1261.
\doi{10.2307/2981683}

Dawid AP (1984). \dQuote{Present Position and Potential Developments: Some
Personal Views: Statistical Theory: The Prequential Approach}, \emph{Journal of
the Royal Statistical Society: Series A (General)}, \bold{147}(2), 278--292.
\doi{10.2307/2981683}

Diebold FX, Gunther TA, Tay AS (1998). \dQuote{Evaluating Density Forecasts
with Applications to Financial Risk Management}. \emph{International Economic
Review}, \bold{39}(4), 863--883. \doi{10.2307/2527342}

Gneiting T, Balabdaoui F, Raftery AE (2007). \dQuote{Probabilistic Forecasts,
Calibration and Sharpness}.  \emph{Journal of the Royal Statistical Society:
Series B (Methodological)}. \bold{69}(2), 243--268.
\doi{10.1111/j.1467-9868.2007.00587.x}
}
\seealso{
\code{\link{pithist}}, \code{\link{procast}}, \code{\link[graphics]{hist}}
}
\keyword{hplot}
