% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pithist.R
\name{plot.pithist}
\alias{plot.pithist}
\alias{lines.pithist}
\alias{autoplot.pithist}
\title{S3 Methods for Plotting PIT Histograms}
\usage{
\method{plot}{pithist}(
  x,
  single_graph = FALSE,
  style = c("histogram", "lines"),
  confint = TRUE,
  ref = TRUE,
  xlim = c(0, 1),
  ylim = c(0, NA),
  xlab = NULL,
  ylab = NULL,
  main = NULL,
  col = "black",
  fill = adjustcolor("black", alpha.f = 0.2),
  border = "black",
  alpha_min = 0.2,
  lwd = NULL,
  lty = 1,
  axes = TRUE,
  box = TRUE,
  ...
)

\method{lines}{pithist}(
  x,
  confint = FALSE,
  ref = FALSE,
  col = "black",
  fill = adjustcolor("black", alpha.f = 0.2),
  alpha_min = 0.2,
  lwd = 2,
  lty = 1,
  ...
)

\method{autoplot}{pithist}(
  object,
  single_graph = FALSE,
  style = c("histogram", "lines"),
  confint = TRUE,
  ref = TRUE,
  xlim = c(0, 1),
  ylim = c(0, NA),
  xlab = NULL,
  ylab = NULL,
  main = NULL,
  colour = "black",
  fill = "darkgray",
  border = "black",
  alpha_min = 0.2,
  size = NULL,
  linetype = 1,
  legend = FALSE,
  ...
)
}
\arguments{
\item{single_graph}{logical. Should all computed extended reliability
diagrams be plotted in a single graph?}

\item{style}{character specifying the style of pithist. For \code{style = "histogram"}
a traditional PIT hisogram is drawn, for \code{style = "lines"} solely the upper border 
line is plotted. For \code{single_graph = TRUE}, always line-style PIT histograms are 
plotted.}

\item{confint}{logical. Should confident intervals be drawn?}

\item{ref, col, fill, border, alpha_min, lwd, lty, axes, box}{additional graphical
parameters for base plots, whereby \code{x} is a object of class \code{pithist}.}

\item{xlim, ylim}{graphical parameters. These may pertain either to the whole
plot or just the histogram or just the fitted line.}

\item{xlab, ylab, main}{graphical parameters.}

\item{\dots}{further graphical parameters.}

\item{object, x}{an object of class \code{\link{pithist}}.}

\item{colour, size, linetype, legend}{graphical parameters passed for 
\code{ggplot2} style plots, whereby \code{object} is a object of class \code{pithist}.}
}
\description{
Generic plotting functions for PIT histograms of the class \code{"pithist"}
computed by \code{link{pithist}}.
}
\details{
PIT histograms graphically evaluate the probability integral transform (PIT),
i.e., the value that the predictive CDF attains at the observation, with a
uniform distribution. For a well calibrated model fit, the observation will be
drawn from the predictive distribution and the PIT will have a standard uniform
distribution. 

PIT histograms can be rendered as \code{ggplot2} or base *R* graphics by using
the generics \code{\link[ggplot2]{autoplot}} or \code{\link[graphics]{plot}}. 
For a single base *R* graphically panel, \code{\link{lines}} adds an additional PIT histogram.
}
\examples{

## speed and stopping distances of cars
m1_lm <- lm(dist ~ speed, data = cars)

## compute and plot pithist
pithist(m1_lm)

## customize colors and style
pithist(m1_lm, ref = "blue", lty = 2, pch = 20, style = "lines")

## add separate model
if (require("crch", quietly = TRUE)) {
  m1_crch <- crch(dist ~ speed | speed, data = cars)
  lines(pithist(m1_crch, plot = FALSE), col = 2, lty = 2, confint = 2)
}

#-------------------------------------------------------------------------------
if (require("crch")) {

  ## precipitation observations and forecasts for Innsbruck
  data("RainIbk", package = "crch")
  RainIbk <- sqrt(RainIbk)
  RainIbk$ensmean <- apply(RainIbk[,grep('^rainfc',names(RainIbk))], 1, mean)
  RainIbk$enssd <- apply(RainIbk[,grep('^rainfc',names(RainIbk))], 1, sd)
  RainIbk <- subset(RainIbk, enssd > 0)

  ## linear model w/ constant variance estimation
  m2_lm <- lm(rain ~ ensmean, data = RainIbk)

  ## logistic censored model 
  m2_crch <- crch(rain ~ ensmean | log(enssd), data = RainIbk, left = 0, dist = "logistic")

  ## compute pithists
  pit2_lm <- pithist(m2_lm, plot = FALSE)
  pit2_crch <- pithist(m2_crch, plot = FALSE)

  ## plot in single graph with style "lines"
  plot(c(pit2_lm, pit2_crch), col = c(1, 2), confint = c(1, 2), ref = 3, 
    style = "lines", single_graph = TRUE)
}

#-------------------------------------------------------------------------------
## determinants for male satellites to nesting horseshoe crabs
data("CrabSatellites", package = "countreg")

## linear poisson model
m3_pois  <- glm(satellites ~ width + color, data = CrabSatellites, family = poisson)

## compute and plot rootogram as ggplot2 graphic
pithist(m3_pois, plot = "ggplot2")

}
\references{
Agresti A, Coull AB (1998). \dQuote{Approximate is Better than ``Exact''
for Interval Estimation of Binomial Proportions.} \emph{The American
Statistician}, \bold{52}(2), 119--126. \doi{10.1080/00031305.1998.10480550}

Czado C, Gneiting T, Held L (2009). \dQuote{Predictive Model
Assessment for Count Data.} \emph{Biometrics}, \bold{65}(4), 1254--1261. 
\doi{10.2307/2981683}
 
Dawid AP (1984). \dQuote{Present Position and Potential Developments: Some
Personal Views: Statistical Theory: The Prequential Approach}, \emph{Journal of
the Royal Statistical Society: Series A (General)}, \bold{147}(2), 278--292.
\doi{10.2307/2981683}

Diebold FX, Gunther TA, Tay AS (1998). \dQuote{Evaluating Density Forecasts
with Applications to Financial Risk Management}. \emph{International Economic
Review}, \bold{39}(4), 863--883. \doi{10.2307/2527342}

Gneiting T, Balabdaoui F, Raftery AE (2007). \dQuote{Probabilistic Forecasts,
Calibration and Sharpness}.  \emph{Journal of the Royal Statistical Society:
Series B (Methodological)}. \bold{69}(2), 243--268.
\doi{10.1111/j.1467-9868.2007.00587.x}
}
\seealso{
\code{\link{pithist}}, \code{\link{procast}}, \code{\link[graphics]{hist}}
}
\keyword{hplot}
