% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqrplot.R
\name{plot.qqrplot}
\alias{plot.qqrplot}
\alias{points.qqrplot}
\alias{autoplot.qqrplot}
\title{Plotting Q-Q Plots for Quantile Residuals}
\usage{
\method{plot}{qqrplot}(
  x,
  single_graph = FALSE,
  confint = TRUE,
  ref = TRUE,
  xlim = c(NA, NA),
  ylim = c(NA, NA),
  xlab = NULL,
  ylab = NULL,
  main = NULL,
  col = adjustcolor("black", alpha.f = 0.4),
  fill = adjustcolor("black", alpha.f = 0.2),
  alpha_min = 0.2,
  pch = 19,
  axes = TRUE,
  box = TRUE,
  ...
)

\method{points}{qqrplot}(
  x,
  confint = FALSE,
  ref = FALSE,
  col = adjustcolor("black", alpha.f = 0.4),
  fill = adjustcolor("black", alpha.f = 0.2),
  alpha_min = 0.2,
  pch = 19,
  ...
)

\method{autoplot}{qqrplot}(
  object,
  single_graph = FALSE,
  confint = TRUE,
  ref = TRUE,
  xlim = c(NA, NA),
  ylim = c(NA, NA),
  xlab = NULL,
  ylab = NULL,
  main = NULL,
  colour = adjustcolor("black", alpha.f = 0.4),
  fill = adjustcolor("black", alpha.f = 0.2),
  alpha_min = 0.2,
  size = 2,
  shape = 19,
  linetype = 1,
  legend = FALSE,
  ...
)
}
\arguments{
\item{x, object}{an object of class \code{qqrplot}.}

\item{single_graph}{logical. Should all computed extended reliability
diagrams be plotted in a single graph?}

\item{confint}{logical or quantile specification. Should the range of
quantiles of the randomized quantile residuals be visualized? If
\code{TRUE}, then \code{range = c(0.01, 0.99)} is used.}

\item{ref, xlim, ylim, col, fill, alpha_min, pch, axes, box}{additional graphical
parameters for base plots, whereby \code{x} is a object of class \code{pithist}.}

\item{xlab, ylab, main, \dots}{graphical plotting parameters passed to
\code{\link[graphics]{plot}} or \code{\link[graphics]{points}},
respectively.}

\item{colour, size, shape, linetype, legend}{graphical parameters passed for 
\code{ggplot2} style plots, whereby \code{object} is a object of class \code{pithist}.}
}
\value{
An list is returned invisibly with: \item{normal}{the theoretical
normal quantiles,} \item{residuals}{the empirical quantile residuals.}
}
\description{
Visualize goodness of fit of regression models by Q-Q plots using quantile
residuals.
}
\details{
Q-Q residual draw quantile residuals (by default: transformed to standard
normal scale) against theoretical quantiles from the same distribution.
Alternatively, transformations to other distributions can also be used,
specifically using no transformation at all, i.e., remaining on the uniform
scale (via \code{trafo = NULL} or equivalently \code{qunif} or
\code{identity}).

Additional options are offered for models with discrete responses where
randomization of quantiles is needed.
}
\examples{

## speed and stopping distances of cars
m1_lm <- lm(dist ~ speed, data = cars)

## compute and plot qqrplot
qqrplot(m1_lm)

## customize colors
qqrplot(m1_lm, ref = "blue", lty = 2, pch = 20)

## add separate model
if (require("crch", quietly = TRUE)) {
  m1_crch <- crch(dist ~ speed | speed, data = cars)
  points(qqrplot(m1_crch, plot = FALSE), col = 2, lty = 2, confint = 2)
}

#-------------------------------------------------------------------------------
if (require("crch")) {

  ## precipitation observations and forecasts for Innsbruck
  data("RainIbk", package = "crch")
  RainIbk <- sqrt(RainIbk)
  RainIbk$ensmean <- apply(RainIbk[,grep('^rainfc',names(RainIbk))], 1, mean)
  RainIbk$enssd <- apply(RainIbk[,grep('^rainfc',names(RainIbk))], 1, sd)
  RainIbk <- subset(RainIbk, enssd > 0)

  ## linear model w/ constant variance estimation
  m2_lm <- lm(rain ~ ensmean, data = RainIbk)

  ## logistic censored model 
  m2_crch <- crch(rain ~ ensmean | log(enssd), data = RainIbk, left = 0, dist = "logistic")

  ## compute qqrplots
  qq2_lm <- qqrplot(m2_lm, plot = FALSE)
  qq2_crch <- qqrplot(m2_crch, plot = FALSE)

  ## plot in single graph
  plot(c(qq2_lm, qq2_crch), col = c(1, 2), confint = c(1, 2), ref = 3, single_graph = TRUE)
}

#-------------------------------------------------------------------------------
## determinants for male satellites to nesting horseshoe crabs
data("CrabSatellites", package = "countreg")

## linear poisson model
m3_pois  <- glm(satellites ~ width + color, data = CrabSatellites, family = poisson)

## compute and plot rootogram as ggplot2 graphic
qqrplot(m3_pois, plot = "ggplot2")

}
\references{
Dunn KP, Smyth GK (1996). \dQuote{Randomized Quantile
Residuals.} \emph{Journal of Computational and Graphical Statistics},
\bold{5}, 1--10.
}
\seealso{
\code{\link{qresiduals}}, \code{\link[stats]{qqnorm}}
}
\keyword{hplot}
