% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rootogram.R
\name{plot.rootogram}
\alias{plot.rootogram}
\alias{autoplot.rootogram}
\title{Plotting Rootograms for Assessing Goodness of Fit of Probability Models}
\usage{
\method{plot}{rootogram}(
  x,
  ref = TRUE,
  xlim = c(NA, NA),
  ylim = c(NA, NA),
  xlab = NULL,
  ylab = NULL,
  main = NULL,
  border = "black",
  fill = adjustcolor("black", alpha.f = 0.2),
  col = 2,
  lwd = 2,
  pch = 19,
  lty = 1,
  type = NULL,
  axes = TRUE,
  box = FALSE,
  ...
)

\method{autoplot}{rootogram}(
  object,
  ref = TRUE,
  xlim = c(NA, NA),
  ylim = c(NA, NA),
  xlab = NULL,
  ylab = NULL,
  main = NULL,
  border = "black",
  fill = "darkgray",
  colour = 2,
  size = 1,
  shape = 19,
  linetype = 1,
  type = NULL,
  legend = FALSE,
  ...
)
}
\arguments{
\item{x, object}{an object of class \code{\link{rootogram}}.}

\item{ref}{logical. Should a reference line be plotted?}

\item{xlim, ylim, border, fill, col, lwd, pch, lty, type, axes, box}{graphical
parameters. These may pertain either to the whole plot or just the histogram
or just the fitted line.}

\item{xlab, ylab, main}{graphical parameters.}

\item{\dots}{further graphical parameters passed to the plotting function.}

\item{colour, size, shape, linetype}{graphical parameters passed to
\code{geom_line} and \code{geom_point}, respectively.}

\item{legend}{logical. Should a legend be added in the \code{ggplot2} style
graphic?}
}
\value{
An object of class \code{"rootogram"} inheriting from
\code{"data.frame"} with the following variables: \item{observed}{observed
frequencies,} \item{expected}{fitted frequencies,} \item{x}{histogram
interval midpoints on the x-axis,} \item{y}{bottom coordinate of the
histogram bars,} \item{width}{widths of the histogram bars,}
\item{height}{height of the histogram bars,} \item{line}{y-coordinates of
the fitted curve.} Additionally, \code{style}, \code{scale}, \code{xlab},
\code{ylab}, and \code{main} are stored as attributes.
}
\description{
Rootograms graphically compare (square roots) of empirical frequencies with
fitted frequencies from a probability model.
}
\details{
Rootograms graphically compare frequencies of empirical distributions and
fitted probability models. For the observed distribution the histogram is
drawn on a square root scale (hence the name) and superimposed with a line
for the fitted frequencies. The histogram can be \code{"standing"} on the
x-axis (as usual), or \code{"hanging"} from the fitted curve, or a
\code{"suspended"} histogram of deviations can be drawn.

\code{rootogram} is the generic function for generating rootograms from data
or fitted model objects. The workhorse function is the default method (that
computes all necessary coordinates based on observed and fitted frequencies
and the breaks for the histogram intervals) and the associating \code{plot}
method that carries out the actual drawing (using base graphics).

There is a wide range of further \code{rootogram} methods that all take the
following approach: based on a fitted probability model observed and
expected frequencies are computed and then the default method is called.
Currently, there is a method for \code{\link[stats]{glm}}.

Furthermore, there is a \code{numeric} method that uses
\code{link[MASS]{fitdistr}} to obtain a fitted (by maximum likelihood)
probability model for a univariate variable.  For this method, \code{fitted}
can either be a character string or a density function that is passed to
\code{fitdistr}. In the latter case, a \code{start} list also has to be
supplied.

In addition to the \code{plot} method for rootogram objects, there are also
two methods that combine two (or more) rootograms: \code{c}/\code{rbind}
creates a set of rootograms that can then be plotted in one go. The \code{+}
method adds up the observed and fitted frequencies from two rootograms (if
these use the same intervals).

The \code{\link[ggplot2]{autoplot}} method creates a
\code{\link[ggplot2]{ggplot}} version of the rootogram.
}
\note{
Note that there is also a \code{\link[vcd]{rootogram}} function in the
\pkg{vcd} package that is similar to the \code{numeric} method provided
here. However, it is much more limited in scope, hence a function has been
created here.
}
\examples{

## speed and stopping distances of cars
m1_lm <- lm(dist ~ speed, data = cars)

## compute and plot rootogram
rootogram(m1_lm)

## customize colors
rootogram(m1_lm, ref = "blue", lty = 2, pch = 20)

## add separate model
if (require("crch", quietly = TRUE)) {
  m1_crch <- crch(dist ~ speed | speed, data = cars)
  points(rootogram(m1_crch, plot = FALSE), col = 2, lty = 2)
}

#-------------------------------------------------------------------------------
if (require("crch")) {

  ## precipitation observations and forecasts for Innsbruck
  data("RainIbk", package = "crch")
  RainIbk <- sqrt(RainIbk)
  RainIbk$ensmean <- apply(RainIbk[,grep('^rainfc',names(RainIbk))], 1, mean)
  RainIbk$enssd <- apply(RainIbk[,grep('^rainfc',names(RainIbk))], 1, sd)
  RainIbk <- subset(RainIbk, enssd > 0)

  ## linear model w/ constant variance estimation
  m2_lm <- lm(rain ~ ensmean, data = RainIbk)

  ## logistic censored model 
  m2_crch <- crch(rain ~ ensmean | log(enssd), data = RainIbk, left = 0, dist = "logistic")

  ## compute rootograms
  r2_lm <- rootogram(m2_lm, plot = FALSE)
  r2_crch <- rootogram(m2_crch, plot = FALSE)

  ## plot in single graph
  plot(c(r2_lm, r2_crch), col = c(1, 2))
}

#-------------------------------------------------------------------------------
## artificial data from negative binomial (mu = 3, theta = 2)
## and Poisson (mu = 3) distribution
set.seed(1090)
y <- rnbinom(100, mu = 3, size = 2)
x <- rpois(100, lambda = 3)

## glm method: fitted values via glm()
m3_pois <- glm(y ~ x, family = poisson)

## correctly specified Poisson model fit
par(mfrow = c(1, 3))
r3a_pois <- rootogram(m3_pois, style = "standing", ylim = c(-2.2, 4.8), main = "Standing")
r3b_pois <- rootogram(m3_pois, style = "hanging", ylim = c(-2.2, 4.8), main = "Hanging")
r3c_pois <- rootogram(m3_pois, style = "suspended", ylim = c(-2.2, 4.8), main = "Suspended")
par(mfrow = c(1, 1))

}
\references{
Friendly M (2000), \emph{Visualizing Categorical Data}. SAS
Institute, Cary.

Kleiber C, Zeileis A (2016).  \dQuote{Visualizing Count Data Regressions
Using Rootograms.} \emph{The American Statistician}, \bold{70}(3), 296--303.
c("\\Sexpr[results=rd,stage=build]{tools:::Rd_expr_doi(\"#1\")}",
"10.1080/00031305.2016.1173590")\Sexpr{tools:::Rd_expr_doi("10.1080/00031305.2016.1173590")}.

Tukey JW (1977). \emph{Exploratory Data Analysis}. Addison-Wesley, Reading.
}
\seealso{
\code{\link[stats]{glm}}
}
\keyword{hplot}
