% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wormplot.R
\name{plot.wormplot}
\alias{plot.wormplot}
\alias{points.wormplot}
\alias{autoplot.wormplot}
\title{Plotting Worm Plots for Quantile Residuals}
\usage{
\method{plot}{wormplot}(
  x,
  single_graph = FALSE,
  confint = TRUE,
  ref = TRUE,
  xlim = c(NA, NA),
  ylim = c(NA, NA),
  xlab = NULL,
  ylab = NULL,
  main = NULL,
  col = adjustcolor("black", alpha.f = 0.4),
  fill = adjustcolor("black", alpha.f = 0.2),
  alpha_min = 0.2,
  pch = 19,
  axes = TRUE,
  box = TRUE,
  ...
)

\method{points}{wormplot}(
  x,
  confint = FALSE,
  ref = FALSE,
  col = "black",
  fill = adjustcolor("black", alpha.f = 0.4),
  alpha_min = 0.2,
  pch = 19,
  ...
)

\method{autoplot}{wormplot}(
  object,
  single_graph = FALSE,
  confint = TRUE,
  ref = TRUE,
  xlim = c(NA, NA),
  ylim = c(NA, NA),
  xlab = NULL,
  ylab = NULL,
  main = NULL,
  colour = adjustcolor("black", alpha.f = 0.4),
  fill = adjustcolor("black", alpha.f = 0.2),
  alpha_min = 0.2,
  size = 2,
  shape = 19,
  linetype = 1,
  legend = FALSE,
  ...
)
}
\arguments{
\item{x, object}{an object of class \code{reliagram}.}

\item{single_graph}{logical. Should all computed extended reliability
diagrams be plotted in a single graph?}

\item{confint}{logical or quantile specification. Should the range of
quantiles of the randomized quantile residuals be visualized? If
\code{TRUE}, then \code{range = c(0.01, 0.99)} is used.}

\item{ref, xlim, ylim, col, fill, alpha_min, pch, axes, box}{additional graphical
parameters for base plots, whereby \code{x} is a object of class \code{pithist}.}

\item{xlab, ylab, main, \dots}{graphical plotting parameters passed to
\code{\link[graphics]{plot}} or \code{\link[graphics]{points}},
respectively.}

\item{colour, size, shape, linetype, legend}{graphical parameters passed for 
\code{ggplot2} style plots, whereby \code{object} is a object of class \code{pithist}.}
}
\value{
An list is returned invisibly with: \item{normal}{the theoretical
normal quantiles,} \item{residuals}{the empirical quantile residuals.}
}
\description{
Visualize goodness of fit of regression models by Worm plots using quantile
residuals.
}
\details{
FIXME: Q-Q residual draw quantile residuals (by default: transformed to
standard normal scale) against theoretical quantiles from the same
distribution. Alternatively, transformations to other distributions can also
be used, specifically using no transformation at all, i.e., remaining on the
uniform scale (via \code{trafo = NULL} or equivalently \code{qunif} or
\code{identity}).

Additional options are offered for models with discrete responses where
randomization of quantiles is needed.
}
\examples{

data("CrabSatellites", package = "countreg")
CrabSatellites2 <- CrabSatellites[CrabSatellites$satellites <= 1, ]

m1 <- glm(satellites ~ width + color, data = CrabSatellites, family = poisson)
m2 <- glm(satellites ~ width + color, data = CrabSatellites2, family = binomial)
m3 <- lm(dist ~ speed, data = cars)

w1 <- wormplot(m1, nsim = 100, confint = TRUE)
w2 <- wormplot(m2, nsim = 100, confint = TRUE, plot = FALSE)
w3 <- wormplot(m3, nsim = 100, confint = TRUE, plot = FALSE)

plot(c(w1, w2), single_graph = FALSE, fill = c(1, 3), ref = c(2, 2))
points(w3, col = "lightblue")

}
\references{
FIXME: Dunn KP, Smyth GK (1996). \dQuote{Randomized Quantile
Residuals.} \emph{Journal of Computational and Graphical Statistics},
\bold{5}, 1--10.
}
\seealso{
\code{\link{qresiduals}}, \code{\link[stats]{qqnorm}}
}
\keyword{hplot}
