% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wormplot.R
\name{plot.wormplot}
\alias{plot.wormplot}
\alias{points.wormplot}
\alias{autoplot.wormplot}
\title{S3 Methods for Plotting Worm Plots}
\usage{
\method{plot}{wormplot}(
  x,
  single_graph = FALSE,
  confint = TRUE,
  ref = TRUE,
  xlim = c(NA, NA),
  ylim = c(NA, NA),
  xlab = NULL,
  ylab = NULL,
  main = NULL,
  col = adjustcolor("black", alpha.f = 0.4),
  fill = adjustcolor("black", alpha.f = 0.2),
  alpha_min = 0.2,
  pch = 19,
  axes = TRUE,
  box = TRUE,
  ...
)

\method{points}{wormplot}(
  x,
  confint = FALSE,
  ref = FALSE,
  col = "black",
  fill = adjustcolor("black", alpha.f = 0.4),
  alpha_min = 0.2,
  pch = 19,
  ...
)

\method{autoplot}{wormplot}(
  object,
  single_graph = FALSE,
  confint = TRUE,
  ref = TRUE,
  xlim = c(NA, NA),
  ylim = c(NA, NA),
  xlab = NULL,
  ylab = NULL,
  main = NULL,
  colour = adjustcolor("black", alpha.f = 0.4),
  fill = adjustcolor("black", alpha.f = 0.2),
  alpha_min = 0.2,
  size = 2,
  shape = 19,
  linetype = 1,
  legend = FALSE,
  ...
)
}
\arguments{
\item{x, object}{an object of class \code{wormplot}.}

\item{single_graph}{logical. Should all computed extended reliability
diagrams be plotted in a single graph?}

\item{confint}{logical or quantile specification. Should the range of
quantiles of the randomized quantile residuals be visualized? If
\code{TRUE}, then \code{range = c(0.01, 0.99)} is used.}

\item{ref, xlim, ylim, col, fill, alpha_min, pch, axes, box}{additional graphical
parameters for base plots, whereby \code{x} is a object of class \code{wormplot}.}

\item{xlab, ylab, main, \dots}{graphical plotting parameters passed to
\code{\link[graphics]{plot}} or \code{\link[graphics]{points}},
respectively.}

\item{colour, size, shape, linetype, legend}{graphical parameters passed for 
\code{ggplot2} style plots, whereby \code{object} is a object of class \code{wormplot}.}
}
\description{
Generic plotting functions for worm plot of the class \code{"wormplot"}
computed by \code{link{wormplot}}.
}
\details{
Worm plots (de-trended Q-Q plots) draw deviations of quantile residuals (by
default: transformed to standard normal scale) and theoretical quantiles from
the same distribution against the same theoretical quantiles.

Worm plots can be rendered as \code{ggplot2} or base R graphics by using
the generics \code{\link[ggplot2]{autoplot}} or \code{\link[graphics]{plot}}. 
For a single base R graphically panel, \code{\link{points}} adds an additional 
worm plot.
}
\examples{

## speed and stopping distances of cars
m1_lm <- lm(dist ~ speed, data = cars)

## compute and plot wormplot
wormplot(m1_lm)

## customize colors
wormplot(m1_lm, ref = "blue", lty = 2, pch = 20)

## add separate model
if (require("crch", quietly = TRUE)) {
  m1_crch <- crch(dist ~ speed | speed, data = cars)
  points(wormplot(m1_crch, plot = FALSE), col = 2, lty = 2, confint = 2)
}

#-------------------------------------------------------------------------------
if (require("crch")) {

  ## precipitation observations and forecasts for Innsbruck
  data("RainIbk", package = "crch")
  RainIbk <- sqrt(RainIbk)
  RainIbk$ensmean <- apply(RainIbk[,grep('^rainfc',names(RainIbk))], 1, mean)
  RainIbk$enssd <- apply(RainIbk[,grep('^rainfc',names(RainIbk))], 1, sd)
  RainIbk <- subset(RainIbk, enssd > 0)

  ## linear model w/ constant variance estimation
  m2_lm <- lm(rain ~ ensmean, data = RainIbk)

  ## logistic censored model 
  m2_crch <- crch(rain ~ ensmean | log(enssd), data = RainIbk, left = 0, dist = "logistic")

  ## compute wormplots
  wp2_lm <- wormplot(m2_lm, plot = FALSE)
  wp2_crch <- wormplot(m2_crch, plot = FALSE)

  ## plot in single graph
  plot(c(wp2_lm, wp2_crch), col = c(1, 2), confint = c(1, 2), ref = 3, single_graph = TRUE)
}

#-------------------------------------------------------------------------------
## determinants for male satellites to nesting horseshoe crabs
data("CrabSatellites", package = "countreg")

## linear poisson model
m3_pois  <- glm(satellites ~ width + color, data = CrabSatellites, family = poisson)

## compute and plot wormplot as "ggplot2" graphic
wormplot(m3_pois, plot = "ggplot2")

}
\references{
Dunn KP, Smyth GK (1996). \dQuote{Randomized Quantile
Residuals.} \emph{Journal of Computational and Graphical Statistics},
\bold{5}, 1--10.
}
\seealso{
\code{\link{wormplot}}, \code{\link{qresiduals}}, \code{\link[stats]{qqnorm}}
}
\keyword{hplot}
