% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqrplot.R
\name{qqrplot}
\alias{qqrplot}
\alias{qqrplot.default}
\alias{plot.qqrplot}
\alias{points.qqrplot}
\alias{c.qqrplot}
\alias{autoplot.qqrplot}
\alias{rbind.qqrplot}
\title{Q-Q Plots for Quantile Residuals}
\usage{
qqrplot(object, ...)

\method{qqrplot}{default}(
  object,
  newdata = NULL,
  plot = TRUE,
  class = NULL,
  trafo = qnorm,
  nsim = 1L,
  delta = NULL,
  confint = TRUE,
  confint_level = 0.95,
  confint_nsim = 250,
  confint_seed = 1,
  single_graph = FALSE,
  xlab = "Theoretical quantiles",
  ylab = "Quantile residuals",
  main = NULL,
  ...
)

\method{c}{qqrplot}(...)

\method{rbind}{qqrplot}(...)

\method{plot}{qqrplot}(
  x,
  single_graph = FALSE,
  confint = TRUE,
  ref = TRUE,
  xlim = c(NA, NA),
  ylim = c(NA, NA),
  xlab = NULL,
  ylab = NULL,
  main = NULL,
  col = adjustcolor("black", alpha.f = 0.4),
  fill = adjustcolor("black", alpha.f = 0.2),
  alpha_min = 0.2,
  pch = 19,
  axes = TRUE,
  box = TRUE,
  ...
)

\method{points}{qqrplot}(
  x,
  confint = FALSE,
  ref = FALSE,
  col = adjustcolor("black", alpha.f = 0.4),
  fill = adjustcolor("black", alpha.f = 0.2),
  alpha_min = 0.2,
  pch = 19,
  ...
)

\method{autoplot}{qqrplot}(
  object,
  single_graph = FALSE,
  confint = TRUE,
  ref = TRUE,
  xlim = c(NA, NA),
  ylim = c(NA, NA),
  xlab = NULL,
  ylab = NULL,
  main = NULL,
  colour = adjustcolor("black", alpha.f = 0.4),
  fill = adjustcolor("black", alpha.f = 0.2),
  alpha_min = 0.2,
  size = 2,
  shape = 19,
  linetype = 1,
  legend = FALSE,
  ...
)
}
\arguments{
\item{object}{an object (for which a \code{\link{qresiduals}} method
exists).}

\item{newdata}{optionally, a data frame in which to look for variables with
which to predict. If omitted, the original observations are used.}

\item{plot}{Should the \code{plot} or \code{autoplot} method be called to
draw the computed Q-Q plot? Either set \code{plot} expicitly to "base" vs.
"ggplot2" to choose the type of plot, or for a logical \code{plot} argument
it's chosen conditional if the package \code{ggplot2} is loaded.}

\item{class}{Should the invisible return value be either a \code{data.frame}
or a \code{tibble}. Either set \code{class} expicitly to "data.frame" vs.
"tibble", or for NULL it's chosen automatically conditional if the package
\code{tibble} is loaded.}

\item{trafo}{function for tranforming residuals from probability scale to a
different distribution scale (default: Gaussian).}

\item{nsim, delta}{arguments passed to \code{qresiduals}.}

\item{confint}{logical or quantile specification. Should the range of
quantiles of the randomized quantile residuals be visualized? If
\code{TRUE}, then \code{range = c(0.01, 0.99)} is used.}

\item{confint_level}{numeric. The confidence level required.}

\item{confint_nsim}{numeric. The number of simulated quantiles.}

\item{confint_seed}{numeric. The seed to be set for calculating the
confidence interval.}

\item{single_graph}{logical. Should all computed extended reliability
diagrams be plotted in a single graph?}

\item{xlab, ylab, main, \dots}{graphical plotting parameters passed to
\code{\link[graphics]{plot}} or \code{\link[graphics]{points}},
respectively.}

\item{x, ref, xlim, ylim, col, fill, alpha_min, pch, axes, box}{additional graphical
parameters for base plots, whereby \code{x} is a object of class \code{pithist}.}

\item{colour, size, shape, linetype, legend}{graphical parameters passed for 
\code{ggplot2} style plots, whereby \code{object} is a object of class \code{pithist}.}
}
\value{
An list is returned invisibly with: \item{normal}{the theoretical
normal quantiles,} \item{residuals}{the empirical quantile residuals.}
}
\description{
Visualize goodness of fit of regression models by Q-Q plots using quantile
residuals.
}
\details{
Q-Q residual draw quantile residuals (by default: transformed to standard
normal scale) against theoretical quantiles from the same distribution.
Alternatively, transformations to other distributions can also be used,
specifically using no transformation at all, i.e., remaining on the uniform
scale (via \code{trafo = NULL} or equivalently \code{qunif} or
\code{identity}).

Additional options are offered for models with discrete responses where
randomization of quantiles is needed.
}
\examples{

data("CrabSatellites", package = "countreg")
CrabSatellites2 <- CrabSatellites[CrabSatellites$satellites <= 1, ]

m1 <- glm(satellites ~ width + color, data = CrabSatellites, family = poisson)
m2 <- glm(satellites ~ width + color, data = CrabSatellites2, family = binomial)
m3 <- lm(dist ~ speed, data = cars)

q1 <- qqrplot(m1, nsim = 100, confint = TRUE)
q2 <- qqrplot(m2, nsim = 100, confint = TRUE, plot = FALSE)
q3 <- qqrplot(m3, nsim = 100, confint = TRUE, plot = FALSE)

plot(c(q1, q2), single_graph = FALSE, fill = c(1, 3), ref = c(2, 2))
points(q3, col = "lightblue")

}
\references{
Dunn KP, Smyth GK (1996). \dQuote{Randomized Quantile
Residuals.} \emph{Journal of Computational and Graphical Statistics},
\bold{5}, 1--10.
}
\seealso{
\code{\link{qresiduals}}, \code{\link[stats]{qqnorm}}
}
\keyword{hplot}
