% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqrplot.R
\name{qqrplot}
\alias{qqrplot}
\alias{qqrplot.default}
\alias{c.qqrplot}
\title{Q-Q Plots for Quantile Residuals}
\usage{
qqrplot(object, ...)

\method{qqrplot}{default}(
  object,
  newdata = NULL,
  plot = TRUE,
  class = NULL,
  detrend = FALSE,
  trafo = qnorm,
  nsim = 1L,
  delta = NULL,
  confint = TRUE,
  range = TRUE,
  range_level = 0.95,
  range_nsim = 250,
  range_seed = 1,
  single_graph = FALSE,
  xlab = "Theoretical quantiles",
  ylab = if (!detrend) "Quantile residuals" else "Deviation",
  main = NULL,
  ...
)
}
\arguments{
\item{object}{an object from which probability integral transforms can be
extracted using the generic function \code{\link{procast}}.}

\item{newdata}{optionally, a data frame in which to look for variables with
which to predict. If omitted, the original observations are used.}

\item{plot}{Should the \code{plot} or \code{autoplot} method be called to
draw the computed Q-Q plot? Either set \code{plot} expicitly to \code{"base"} vs.
\code{"ggplot2"} to choose the type of plot, or for a logical \code{plot} argument
it's chosen conditional if the package \code{ggplot2} is loaded.}

\item{class}{Should the invisible return value be either a \code{data.frame}
or a \code{tibble}. Either set \code{class} expicitly to \code{"data.frame"} vs.
\code{"tibble"}, or for \code{NULL} it's chosen automatically conditional if the package
\code{tibble} is loaded.}

\item{detrend}{logical. Should the qqrplot be detrended, i.e, plotted as a `wormplot()`?}

\item{trafo}{function for tranforming residuals from probability scale to a
different distribution scale (default: Gaussian).}

\item{nsim, delta}{arguments passed to \code{qresiduals}.}

\item{confint}{logical or character string describing the type for plotting `c("polygon", "line")`.
If not set to `FALSE`, the pointwise confidence interval of the (randomized)
quantile residuals are visualized.}

\item{range}{logical or quantile specification. In case of discrete distributions, should the range 
(confidence interval) of values due to the randomization be visualized? If \code{TRUE}, 
then \code{range = c(0.01, 0.99)} is used.}

\item{range_level}{numeric. The confidence level required for calculating the range of values 
due to the randomization.}

\item{range_nsim}{numeric. The number of simulated quantiles for calculating the range of values
due to the randomization.}

\item{range_seed}{numeric. The seed to be set for calculating the range of values
due to the randomization.}

\item{single_graph}{logical. Should all computed extended reliability
diagrams be plotted in a single graph?}

\item{xlab, ylab, main, \dots}{graphical parameters passed to
\code{\link{plot.qqrplot}} or \code{\link{autoplot.qqrplot}}.}
}
\value{
An object of class \code{"qqrplot"} inheriting from
\code{"data.frame"} or \code{"tibble"} conditional on the argument \code{class}
with the following variables: \item{x}{theoretical quantiles,}
\item{y}{deviations between theoretical and empirical quantiles.} In case of
randomized residuals, \code{nsim} different \code{x} and \code{y} values, and
lower and upper confidence interval bounds (\code{x_rg_lwr}, \code{y_rg_lwr},
\code{x_rg_upr}, \code{y_rg_upr}) can optionally be returned.  Additionally,
\code{xlab}, \code{ylab}, \code{main}, and \code{range_level}, as well as the
trafo function (\code{trafo}) and wether a \code{detrended} Q-Q residuals plot
was computed are stored as attributes.
}
\description{
Visualize goodness of fit of regression models by Q-Q plots using quantile
residuals. If \code{plot = TRUE}, the resulting object of class
\code{"qqrplot"} is plotted by \code{\link{plot.qqrplot}} or
\code{\link{autoplot.qqrplot}} before it is returned, depending on whether the
package \code{ggplot2} is loaded.
}
\details{
Q-Q residuals plots draw quantile residuals (by default: transformed to standard
normal scale) against theoretical quantiles from the same distribution.
Alternatively, transformations to other distributions can also be used,
specifically using no transformation at all, i.e., remaining on the uniform
scale (via \code{trafo = NULL} or equivalently \code{qunif} or
\code{identity}). For computation, \code{\link{qqrplot}} leverages the function
\code{\link{qresiduals}} employing the \code{\link{procast}} generic.

Additional options are offered for models with discrete responses where
randomization of quantiles is needed.

In addition to the \code{plot} and \code{\link[ggplot2]{autoplot}} method for
qqrplot objects, it is also possible to combine two (or more) Q-Q residuals plots by
\code{c}/\code{rbind}, which creates a set of Q-Q residuals plots that can then be
plotted in one go.
}
\examples{

## speed and stopping distances of cars
m1_lm <- lm(dist ~ speed, data = cars)

## compute and plot qqrplot
qqrplot(m1_lm)

#-------------------------------------------------------------------------------
## determinants for male satellites to nesting horseshoe crabs
data("CrabSatellites", package = "countreg")

## linear poisson model
m1_pois <- glm(satellites ~ width + color, data = CrabSatellites, family = poisson)
m2_pois <- glm(satellites ~ color, data = CrabSatellites, family = poisson)

## compute and plot qqrplot as base graphic
q1 <- qqrplot(m1_pois, plot = FALSE)
q2 <- qqrplot(m2_pois, plot = FALSE)

## plot combined qqrplot as "ggplot2" graphic
ggplot2::autoplot(c(q1, q2), single_graph = TRUE, col = c(1, 2), fill = c(1, 2))

}
\references{
Dunn KP, Smyth GK (1996). \dQuote{Randomized Quantile
Residuals.} \emph{Journal of Computational and Graphical Statistics},
\bold{5}, 1--10. \doi{10.2307/1390802}
}
\seealso{
\code{\link{plot.qqrplot}}, \code{\link{wormplot}},
\code{\link{qresiduals}}, \code{\link[stats]{qqnorm}}
}
\keyword{hplot}
