% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqrplot.R
\name{qqrplot}
\alias{qqrplot}
\alias{qqrplot.default}
\alias{c.qqrplot}
\title{Q-Q Plots for Quantile Residuals}
\usage{
qqrplot(object, ...)

\method{qqrplot}{default}(
  object,
  newdata = NULL,
  plot = TRUE,
  class = NULL,
  detrend = FALSE,
  scale = c("normal", "uniform"),
  nsim = 1L,
  delta = NULL,
  simint = TRUE,
  simint_level = 0.95,
  simint_nrep = 250,
  confint = TRUE,
  ref = TRUE,
  xlab = "Theoretical quantiles",
  ylab = if (!detrend) "Quantile residuals" else "Deviation",
  main = NULL,
  ...
)
}
\arguments{
\item{object}{an object from which probability integral transforms can be
extracted using the generic function \code{\link{procast}}.}

\item{newdata}{an optional data frame in which to look for variables with
which to predict. If omitted, the original observations are used.}

\item{plot}{logical or character. Should the \code{plot} or \code{autoplot}
method be called to draw the computed Q-Q plot? Logical \code{FALSE} will
suppress plotting, \code{TRUE} (default) will choose the type of plot
conditional if the package \code{ggplot2} is loaded.  Alternatively
\code{"base"} or \code{"ggplot2"} can be specified to explicitly choose the
type of plot.}

\item{class}{should the invisible return value be either a \code{data.frame}
or a \code{tibble}. Either set \code{class} expicitly to \code{"data.frame"} vs.
\code{"tibble"}, or for \code{NULL} it's chosen automatically conditional if the package
\code{tibble} is loaded.}

\item{detrend}{logical, defaults to \code{FALSE}.
Should the qqrplot be detrended, i.e, plotted as a \code{\link{wormplot}}?}

\item{scale}{character. On which scale should the quantile residuals be
shown: on the probability scale (\code{"uniform"}) or on the normal scale (\code{"normal"}).}

\item{nsim, delta}{arguments passed to \code{\link{proresiduals}}.}

\item{simint}{logical. In case of discrete distributions, should the simulation
(confidence) interval due to the randomization be visualized?}

\item{simint_level}{numeric. The confidence level required for calculating
the simulation (confidence) interval due to the randomization.}

\item{simint_nrep}{numeric (positive; default \code{250}). The number of
repetitions of simulated quantiles for calculating the simulation (confidence)
interval due to the randomization.}

\item{confint}{logical or character describing the style for plotting
confidence intervals. \code{TRUE} (default) and \code{"line"} will add
point-wise confidence intervals of the (randomized) quantile residuals as
lines, \code{"polygon"} will draw a polygon instead, and \code{FALSE}
suppresses the drawing.}

\item{ref}{logical, defaults to \code{TRUE}. Should a reference line be plotted?}

\item{xlab, ylab, main, \dots}{graphical parameters passed to
\code{\link{plot.qqrplot}} or \code{\link{autoplot.qqrplot}}.}
}
\value{
An object of class \code{"qqrplot"} inheriting from
\code{"data.frame"} or \code{"tibble"} conditional on the argument \code{class}
with the following variables:
\item{observed}{deviations between theoretical and empirical quantiles,}
\item{expected}{theoretical quantiles,}
\item{simint_observed_lwr}{lower bound of the simulated confidence interval,}
\item{simint_observed_upr}{upper bound of the simulated confidence interval,}
\item{simint_expected}{TODO: (ML) Description missing.}

In case of \code{nsim > 1}, a set of \code{nsim} pairs of observed and
expected quantiles are returned (\code{observed_1}, \code{expected_1}, ...
\code{observed_nsim}, \code{observed_nsim}) is returned.

The \code{"qqrplot"} also contains additional attributes
\code{xlab}, \code{ylab}, \code{main}, \code{simint_level}, \code{scale},
and \code{detrended} used to create the plot.
}
\description{
Visualize goodness of fit of regression models by Quantile-Quantile (Q-Q) plots using quantile
residuals. If \code{plot = TRUE}, the resulting object of class
\code{"qqrplot"} is plotted by \code{\link{plot.qqrplot}} or
\code{\link{autoplot.qqrplot}} before it is returned, depending on whether the
package \code{ggplot2} is loaded.
}
\details{
Q-Q residuals plots draw quantile residuals (by default on the standard normal
scale) against theoretical quantiles from the same distribution.
Alternatively, quantile residuals can also be compared on the uniform scale
(\code{scale = "uniform"}) using no transformation.  For computation,
\code{\link{qqrplot}} leverages the function \code{\link{proresiduals}} employing
the \code{\link{procast}} generic.

Additional options are offered for models with discrete responses where
randomization of quantiles is needed.

In addition to the \code{plot} and \code{\link[ggplot2]{autoplot}} method for
qqrplot objects, it is also possible to combine two (or more) Q-Q residuals plots by
\code{c}/\code{rbind}, which creates a set of Q-Q residuals plots that can then be
plotted in one go.
}
\examples{

## speed and stopping distances of cars
m1_lm <- lm(dist ~ speed, data = cars)

## compute and plot qqrplot
qqrplot(m1_lm)

#-------------------------------------------------------------------------------
## determinants for male satellites to nesting horseshoe crabs
data("CrabSatellites", package = "countreg")

## linear poisson model
m1_pois <- glm(satellites ~ width + color, data = CrabSatellites, family = poisson)
m2_pois <- glm(satellites ~ color, data = CrabSatellites, family = poisson)

## compute and plot qqrplot as base graphic
q1 <- qqrplot(m1_pois, plot = FALSE)
q2 <- qqrplot(m2_pois, plot = FALSE)

## plot combined qqrplot as "ggplot2" graphic
ggplot2::autoplot(c(q1, q2), single_graph = TRUE, col = c(1, 2), fill = c(1, 2))

## Use different `scale`s with confidence intervals
qqrplot(m1_pois, scale = "uniform")
qqrplot(m1_pois, scale = "normal")
qqrplot(m1_pois, detrend = TRUE, scale = "uniform", confint = "line")
qqrplot(m1_pois, detrend = TRUE, scale = "normal", confint = "line")

}
\references{
Dunn KP, Smyth GK (1996). \dQuote{Randomized Quantile
Residuals.} \emph{Journal of Computational and Graphical Statistics},
\bold{5}(3), 236--244. \doi{10.2307/1390802}
}
\seealso{
\code{\link{plot.qqrplot}}, \code{\link{wormplot}},
\code{\link{proresiduals}}, \code{\link[stats]{qqnorm}}
}
\keyword{hplot}
