% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqrplot.R
\name{qqrplot}
\alias{qqrplot}
\alias{qqrplot.default}
\alias{c.qqrplot}
\title{Q-Q Plots for Quantile Residuals}
\usage{
qqrplot(object, ...)

\method{qqrplot}{default}(
  object,
  newdata = NULL,
  plot = TRUE,
  class = NULL,
  trafo = qnorm,
  nsim = 1L,
  delta = NULL,
  confint = TRUE,
  confint_level = 0.95,
  confint_nsim = 250,
  confint_seed = 1,
  single_graph = FALSE,
  xlab = "Theoretical quantiles",
  ylab = "Quantile residuals",
  main = NULL,
  ...
)
}
\arguments{
\item{object}{an object from which probability integral transforms can be
extracted using the generic function \code{\link{procast}}.}

\item{newdata}{optionally, a data frame in which to look for variables with
which to predict. If omitted, the original observations are used.}

\item{plot}{Should the \code{plot} or \code{autoplot} method be called to
draw the computed Q-Q plot? Either set \code{plot} expicitly to "base" vs.
"ggplot2" to choose the type of plot, or for a logical \code{plot} argument
it's chosen conditional if the package \code{ggplot2} is loaded.}

\item{class}{Should the invisible return value be either a \code{data.frame}
or a \code{tibble}. Either set \code{class} expicitly to "data.frame" vs.
"tibble", or for NULL it's chosen automatically conditional if the package
\code{tibble} is loaded.}

\item{trafo}{function for tranforming residuals from probability scale to a
different distribution scale (default: Gaussian).}

\item{nsim, delta}{arguments passed to \code{qresiduals}.}

\item{confint}{logical or quantile specification. Should the range of
quantiles of the randomized quantile residuals be visualized? If
\code{TRUE}, then \code{range = c(0.01, 0.99)} is used.}

\item{confint_level}{numeric. The confidence level required.}

\item{confint_nsim}{numeric. The number of simulated quantiles.}

\item{confint_seed}{numeric. The seed to be set for calculating the
confidence interval.}

\item{single_graph}{logical. Should all computed extended reliability
diagrams be plotted in a single graph?}

\item{xlab, ylab, main, \dots}{graphical parameters handed passed to
\code{\link{plot.qqrplot}} or \code{\link{autoplot.qqrplot}}.}
}
\value{
An object of class \code{"qqrplot"} inheriting from
\code{"data.frame"} or \code{"tibble"} conditional on the argument \code{class} 
with the following variables: \item{x}{theoretical
quantiles,} \item{y}{empirical quantile residuals.} In case of randomized residuals, \code{nsim}
different \code{x} and \code{y}, and a lower and upper confidence interval bound is returned.
Additionally, \code{xlab}, \code{ylab}, \code{main}, and \code{confint_level} 
are stored as attributes.
}
\description{
Visualize goodness of fit of regression models by Q-Q plots using quantile
residuals. If \code{plot = TRUE}, the resulting object of
class \code{"qqrplot"} is plotted by \code{\link{plot.qqrplot}} or
\code{\link{autoplot.qqrplot}} conditional if the package
\code{ggplot2} is loaded, before it is returned.
}
\details{
Q-Q residuals plots draw quantile residuals (by default: transformed to standard
normal scale) against theoretical quantiles from the same distribution.
Alternatively, transformations to other distributions can also be used,
specifically using no transformation at all, i.e., remaining on the uniform
scale (via \code{trafo = NULL} or equivalently \code{qunif} or
\code{identity}). For computation, \code{\link{qqrplot}} leverages the function
\code{\link{qresiduals}} employing the \code{\link{procast}}.

Additional options are offered for models with discrete responses where
randomization of quantiles is needed.

In addition to the \code{plot} and \code{\link[ggplot2]{autoplot}} method for
qqrplot objects, it is also possible to combine two (or more) Q-Q residuals plots by
\code{c}/\code{rbind}, which creates a set of Q-Q residuals plots that can then be
plotted in one go.
}
\examples{

data("CrabSatellites", package = "countreg")
CrabSatellites2 <- CrabSatellites[CrabSatellites$satellites <= 1, ]

m1 <- glm(satellites ~ width + color, data = CrabSatellites, family = poisson)
m2 <- glm(satellites ~ width + color, data = CrabSatellites2, family = binomial)
m3 <- lm(dist ~ speed, data = cars)

q1 <- qqrplot(m1, nsim = 100, confint = TRUE)
q2 <- qqrplot(m2, nsim = 100, confint = TRUE, plot = FALSE)
q3 <- qqrplot(m3, nsim = 100, confint = TRUE, plot = FALSE)

plot(c(q1, q2), single_graph = FALSE, fill = c(1, 3), ref = c(2, 2))
points(q3, col = "lightblue")

}
\references{
Dunn KP, Smyth GK (1996). \dQuote{Randomized Quantile
Residuals.} \emph{Journal of Computational and Graphical Statistics},
\bold{5}, 1--10.
}
\seealso{
\code{\link{plot.qqrplot}}, \code{\link{qresiduals}}, \code{\link[stats]{qqnorm}}
}
\keyword{hplot}
