% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliagram.R
\name{reliagram}
\alias{reliagram}
\alias{reliagram.default}
\alias{c.reliagram}
\title{Reliagram (Extended Reliability Diagram)}
\usage{
reliagram(object, ...)

\method{reliagram}{default}(
  object,
  newdata = NULL,
  plot = TRUE,
  class = NULL,
  breaks = seq(0, 1, by = 0.1),
  quantiles = 0.5,
  thresholds = NULL,
  confint = TRUE,
  confint_level = 0.95,
  confint_nboot = 250,
  confint_seed = 1,
  single_graph = FALSE,
  xlab = "Forecast probability",
  ylab = "Observed relative frequency",
  main = NULL,
  ...
)
}
\arguments{
\item{object}{an object from which an extended reliability diagram can be
extracted with \code{\link{procast}}.}

\item{\dots}{further graphical parameters.}

\item{newdata}{optionally, a data frame in which to look for variables with
which to predict. If omitted, the original observations are used.}

\item{plot}{Should the \code{plot} or \code{autoplot} method be called to
draw the computed extended reliability diagram? Either set \code{plot}
expicitly to "base" vs. "ggplot2" to choose the type of plot, or for a
logical \code{plot} argument it's chosen conditional if the package
\code{ggplot2} is loaded.}

\item{class}{Should the invisible return value be either a \code{data.frame}
or a \code{tibble}. Either set \code{class} expicitly to "data.frame" vs.
"tibble", or for NULL it's chosen automatically conditional if the package
\code{tibble} is loaded.}

\item{breaks}{numeric vector passed on to \code{\link{cut}} in order to bin
the observations and the predicted probabilities or a function applied to
the predicted probabilities to calculate a numeric value for
\code{\link{cut}}. Typically quantiles to ensure equal number of predictions
per bin, e.g., by \code{breaks = function(x) quantile(x)}.}

\item{quantiles}{numeric vector of quantile probabilities with values in
[0,1] to calculate single or several thresholds. Only used if
\code{thresholds} is not specified. For binary responses typically the
50\%-quantile is used.}

\item{thresholds}{numeric vector specifying both where to cut the
observations into binary values and at which values the predicted
probabilities should be calculated (\code{\link{procast}}).}

\item{confint}{logical. Should confident intervals be calculated and drawn?}

\item{confint_level}{numeric. The confidence level required.}

\item{confint_nboot}{numeric. The number of bootstrap steps.}

\item{confint_seed}{numeric. The seed to be set for the bootstrapping.}

\item{single_graph}{logical. Should all computed extended reliability
diagrams be plotted in a single graph?}

\item{xlab, ylab, main}{graphical parameters.}
}
\description{
Reliagram (extended reliability diagram) for assessing the reliability of a
fitted probabilistic distributional forecast.
}
\details{
TODO: (ML)
}
\examples{

require("crch")
m1 <- lm(dist ~ speed, data = cars)
m2 <- crch(dist ~ speed | speed, data = cars)
m3 <- crch(dist ~ speed | speed, left = 30, data = cars)

rel1 <- reliagram(m1, quantiles = c(0.2, 0.6))
rel2 <- reliagram(m2, plot = FALSE)
rel3 <- reliagram(m3, plot = FALSE)

plot(c(rel1, rel2), single_graph = TRUE, col = c(1, 2, 3), lty = c(1, 2, 3), pch = c(1, 2, 3))

plot(rel1, single_graph = TRUE, col = c(1, 2), fill = c(1, 2))
lines(rel3, col = 3, lty = 2, confint = 3)

reliagram(m1, quantiles = c(0.2, 0.6), minimum = c(2, 6), plot = "ggplot2")

}
\references{
Br\''ocker J, Smith L (2007). \dQuote{Increasing the Reliability
of Reliability Diagrams}.  \emph{Weather and Forecasting}, \bold{22}(3),
651--661. doi:10.1175/WAF993.1.
}
\seealso{
\code{\link{procast}}
}
