% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rootogram.R
\name{rootogram}
\alias{rootogram}
\alias{rootogram.default}
\alias{c.rootogram}
\alias{rbind.rootogram}
\title{Rootograms for Assessing Goodness of Fit of Probability Models}
\usage{
rootogram(object, ...)

\method{rootogram}{default}(
  object,
  newdata = NULL,
  plot = TRUE,
  class = NULL,
  response_type = NULL,
  breaks = NULL,
  width = NULL,
  style = c("hanging", "standing", "suspended"),
  scale = c("sqrt", "raw"),
  expected = TRUE,
  ref = TRUE,
  xlab = NULL,
  ylab = NULL,
  main = NULL,
  ...
)
}
\arguments{
\item{object}{an object from which an rootogram can be extracted with
\code{\link{procast}}.}

\item{\dots}{further graphical parameters passed to the plotting function.}

\item{newdata}{optionally, a data frame in which to look for variables with
which to predict. If omitted, the original observations are used.}

\item{plot}{Should the \code{plot} or \code{autoplot} method be called to
draw the computed extended reliability diagram? Either set \code{plot}
expicitly to \code{"base"} vs. \code{"ggplot2"} to choose the type of plot, or for a
logical \code{plot} argument it's chosen conditional if the package
\code{ggplot2} is loaded.}

\item{class}{Should the invisible return value be either a \code{data.frame}
or a \code{tibble}. Either set \code{class} expicitly to \code{"data.frame"} vs.
\code{"tibble"}, or for \code{NULL} it's chosen automatically conditional if the package
\code{tibble} is loaded.}

\item{response_type}{character. To set the default values for \code{breaks} and
\code{widths}.  Currently different defaults are available for \code{"discrete"}
and \code{"continous"} response distribution, as well as for the special case of a
\code{"logseries"} response.}

\item{breaks}{numeric. Breaks for the histogram intervals.}

\item{width}{numeric. Widths of the histogram bars.}

\item{style}{character specifying the syle of rootogram (see below).}

\item{scale}{character specifying whether raw frequencies or their square
roots (default) should be drawn.}

\item{expected}{Should the expected (fitted) frequencies be plotted? Either logical or as character string defining one of `"both"`, `"line"` or `"point"`.}

\item{ref}{logical. Should a reference line be plotted?}

\item{xlab, ylab, main}{graphical parameters.}
}
\value{
An object of class \code{"rootogram"} inheriting from
\code{"data.frame"} or \code{"tibble"} conditional on the argument \code{class}
with the following variables: \item{observed}{observed
frequencies,} \item{expected}{expected (fitted) frequencies,} \item{x}{histogram
interval midpoints on the x-axis,} \item{y}{bottom coordinate of the
histogram bars,} \item{width}{widths of the histogram bars,}
\item{height}{height of the histogram bars,} \item{line}{y-coordinates of
the fitted curve.} Additionally, \code{style}, \code{scale}, \code{xlab},
\code{ylab}, and \code{main} are stored as attributes.
}
\description{
Rootograms graphically compare (square roots) of empirical frequencies with
expected (fitted) frequencies from a probability model. If \code{plot = TRUE}, the
resulting object of class \code{"rootogram"} is plotted by
\code{\link{plot.rootogram}} or \code{\link{autoplot.rootogram}} before it is
returned, depending on whether the package \code{ggplot2} is loaded.
}
\details{
Rootograms graphically compare frequencies of empirical distributions and
expected (fitted) probability models. For the observed distribution the histogram is
drawn on a square root scale (hence the name) and superimposed with a line
for the expected frequencies. The histogram can be \code{"standing"} on the
x-axis (as usual), or \code{"hanging"} from the expected curve, or a
\code{"suspended"} histogram of deviations can be drawn.

The function \code{\link{rootogram}} leverages the \code{\link{procast}}
generic in order to compute all necessary coordinates based on observed and
expected (fitted) frequencies.

In addition to the \code{plot} and \code{\link[ggplot2]{autoplot}} method for
rootogram objects, it is also possible to combine two (or more) rootograms by
\code{c}/\code{rbind}, which creates a set of rootograms that can then be
plotted in one go.
}
\note{
Note that there is also a \code{\link[vcd]{rootogram}} function in the
\pkg{vcd} package that is similar to the \code{numeric} method provided
here. However, it is much more limited in scope, hence a function has been
created here.
}
\examples{

## plots and output

## number of deaths by horsekicks in Prussian army (Von Bortkiewicz 1898)
deaths <- rep(0:4, c(109, 65, 22, 3, 1))

## fit glm model
m1_pois <- glm(deaths ~ 1, family = poisson)
rootogram(m1_pois)

## inspect output (without plotting)
r1 <- rootogram(m1_pois, plot = FALSE)
r1

## combine plots
plot(c(r1, r1), col = c(1, 2), ref = 4, lty = c(1, 2))


#-------------------------------------------------------------------------------
## different styles

## artificial data from negative binomial (mu = 3, theta = 2)
## and Poisson (mu = 3) distribution
set.seed(1090)
y <- rnbinom(100, mu = 3, size = 2)
x <- rpois(100, lambda = 3)

## glm method: fitted values via glm()
m2_pois <- glm(y ~ x, family = poisson)

## correctly specified Poisson model fit
par(mfrow = c(1, 3))
r1 <- rootogram(m2_pois, style = "standing", ylim = c(-2.2, 4.8), main = "Standing")
r2 <- rootogram(m2_pois, style = "hanging", ylim = c(-2.2, 4.8), main = "Hanging")
r3 <- rootogram(m2_pois, style = "suspended", ylim = c(-2.2, 4.8), main = "Suspended")
par(mfrow = c(1, 1))

#-------------------------------------------------------------------------------
## linear regression with normal/Gaussian response: anorexia data

data("anorexia", package = "MASS")

m3_gauss <- glm(Postwt ~ Prewt + Treat + offset(Prewt), family = gaussian, data = anorexia)

## plot rootogram as "ggplot2" graphic
rootogram(m3_gauss, plot = "ggplot2")
}
\references{
Friendly M (2000), \emph{Visualizing Categorical Data}. SAS
Institute, Cary.

Kleiber C, Zeileis A (2016).  \dQuote{Visualizing Count Data Regressions
Using Rootograms.} \emph{The American Statistician}, \bold{70}(3), 296--303.
\doi{10.1080/00031305.2016.1173590}.

Tukey JW (1977). \emph{Exploratory Data Analysis}. Addison-Wesley, Reading.
}
\seealso{
\code{\link{plot.rootogram}}, \code{\link{procast}}
}
\keyword{hplot}
