% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wormplot.R
\name{wormplot}
\alias{wormplot}
\alias{wormplot.default}
\alias{c.wormplot}
\title{Worm Plots for Quantile Residuals}
\usage{
wormplot(object, ...)

\method{wormplot}{default}(
  object,
  newdata = NULL,
  plot = TRUE,
  class = NULL,
  trafo = qnorm,
  nsim = 1L,
  delta = NULL,
  confint = TRUE,
  confint_level = 0.95,
  confint_nsim = 250,
  confint_seed = 1,
  single_graph = FALSE,
  xlab = "Theoretical quantiles",
  ylab = "Deviation",
  main = NULL,
  ...
)
}
\arguments{
\item{object}{an object from which probability integral transforms can be
extracted using the generic function \code{\link{procast}}.}

\item{newdata}{optionally, a data frame in which to look for variables with
which to predict. If omitted, the original observations are used.}

\item{plot}{Should the \code{plot} or \code{autoplot} method be called to
draw the computed Q-Q plot? Either set \code{plot} expicitly to \code{"base"} vs.
\code{"ggplot2"} to choose the type of plot, or for a logical \code{plot} argument
it's chosen conditional if the package \code{ggplot2} is loaded.}

\item{class}{Should the invisible return value be either a \code{data.frame}
or a \code{tibble}. Either set \code{class} expicitly to \code{"data.frame"} vs.
\code{"tibble"}, or for NULL it's chosen automatically conditional if the package
\code{tibble} is loaded.}

\item{trafo}{function for tranforming residuals from probability scale to a
different distribution scale (default: Gaussian).}

\item{nsim, delta}{arguments passed to \code{qresiduals}.}

\item{confint}{logical or quantile specification. Should the range of
quantiles of the randomized quantile residuals be visualized? If
\code{TRUE}, then \code{range = c(0.01, 0.99)} is used.}

\item{confint_level}{numeric. The confidence level required.}

\item{confint_nsim}{numeric. The number of simulated quantiles.}

\item{confint_seed}{numeric. The seed to be set for calculating the
confidence interval.}

\item{single_graph}{logical. Should all computed extended reliability
diagrams be plotted in a single graph?}

\item{xlab, ylab, main, \dots}{graphical parameters handed passed to
\code{\link{plot.wormplot}} or \code{\link{autoplot.wormplot}}.}
}
\value{
An object of class \code{"wormplot"} inheriting from
\code{"data.frame"} or \code{"tibble"} conditional on the argument \code{class} 
with the following variables: \item{x}{theoretical
quantiles,} \item{y}{deviations of theoretical and empirical quantiles.} 
In case of randomized residuals, \code{nsim}
different \code{x} and \code{y}, and a lower and upper confidence interval bound is returned.
Additionally, \code{xlab}, \code{ylab}, \code{main}, and \code{confint_level}, as well as the
reference function (\code{ref_fun}) are stored as attributes.
}
\description{
Visualize goodness of fit of regression models by worm plots using quantile
residuals. If \code{plot = TRUE}, the resulting object of class
\code{"pithist"} is plotted by \code{\link{plot.pithist}} or
\code{\link{autoplot.pithist}} before it is returned, depending on whether the
package \code{ggplot2} is loaded.
}
\details{
Worm plots (de-trended Q-Q plots) draw deviations of quantile residuals (by
default: transformed to standard normal scale) and theoretical quantiles from
the same distribution against the same theoretical quantiles. For computation,
\code{\link{wormplot}} leverages the function \code{\link{qresiduals}}
employing the \code{\link{procast}}.

Additional options are offered for models with discrete responses where
randomization of quantiles is needed.

In addition to the \code{plot} and \code{\link[ggplot2]{autoplot}} method for
wormplot objects, it is also possible to combine two (or more) worm plots by
\code{c}/\code{rbind}, which creates a set of worm plots that can then be
plotted in one go.
}
\examples{

## speed and stopping distances of cars
m1_lm <- lm(dist ~ speed, data = cars)

## compute and plot wormplot
wormplot(m1_lm)

#-------------------------------------------------------------------------------
## determinants for male satellites to nesting horseshoe crabs
data("CrabSatellites", package = "countreg")

## linear poisson model
m1_pois <- glm(satellites ~ width + color, data = CrabSatellites, family = poisson)
m2_pois <- glm(satellites ~ color, data = CrabSatellites, family = poisson)

## compute and plot wormplot as base graphic
w1 <- wormplot(m1_pois, plot = FALSE)
w2 <- wormplot(m2_pois, plot = FALSE)

## plot combined wormplot as "ggplot2" graphic
ggplot2::autoplot(c(w1, w2), single_graph = TRUE, col = c(1, 2), fill = c(1, 2))

}
\references{
van Buuren S and Fredriks M (2001). \dQuote{Worm plot: simple diagnostic
device for modelling growth reference curves}. \emph{Statistics in
Medicine}, \bold{20}, 1259--1277. \doi{10.1002/sim.746}
}
\seealso{
\code{\link{plot.wormplot}}, \code{\link{qqrplot}}, 
\code{\link{qresiduals}}, \code{\link[stats]{qqnorm}}
}
\keyword{hplot}
